### 2021-09-21 #3746 @whedon
[PRE REVIEW]: Frites: A Python package for functional connectivity analysis and group-level statistics of neurophysiological data
**Submitting author:** @EtienneCmb (<a href="http://orcid.org/0000-0002-7362-3247">Etienne Combrisson</a>)
**Repository:** <a href="https://github.com/brainets/frites" target ="_blank">https://github.com/brainets/frites</a>
**Version:** v0.4.0
