### 2021-10-01 #3746 @meg-simula
:wave: @EtienneCmb I've reviewed the paper and code, and this looks very much ready for review in JOSS. I'll reach out to reviewers now.
