### 2021-10-02 #3746 @madvn
Hi @meg-simula , this work from @EtienneCmb et. al. looks quite interesting. While I'd be happy to review it, I'd like to first ask you what timeline you are looking at before I commit to it.
