### 2021-10-03 #3746 @zen-juen
Thanks for reaching out! However, I'll pass since the specificities of the methods in Frites seem to be out of my area of expertise!
