### 2021-10-11 e47559c
Function to compute correlation on FCD
