### 2021-10-11 2001f0c
Implementation of the conn_fcd_corr + test + doc
