+++ docs/source/api/api_connectivity.rst
@@ -22,6 +22,9 @@ Connectivity metrics
 Utility functions
 +++++++++++++++++
 
+Reshaping connectivity outputs
+==============================
+
 .. autosummary::
    :toctree: generated/
 
@@ -29,5 +32,20 @@ Utility functions
    conn_reshape_directed
    conn_ravel_directed
    conn_get_pairs
+
+Metrics to apply on FC
+======================
+
+.. autosummary::
+   :toctree: generated/
+
+   conn_fcd_corr
+
+Define sliding windows
+======================
+
+.. autosummary::
+   :toctree: generated/
+
    define_windows
    plot_windows
