### 2021-10-12 48ef0a0
acf function for computing the auto-correlation
