+++ frites/utils/preproc.py
@@ -7,6 +7,13 @@ from scipy.signal import fftconvolve
 from frites.io import set_log_level, logger
 
 
+###############################################################################
+###############################################################################
+#                          TIME-SERIES PROCESSING
+###############################################################################
+###############################################################################
+
+
 def savgol_filter(x, h_freq, axis=None, sfreq=None, polyorder=5, verbose=None):
     """Filter the data using Savitzky-Golay polynomial method.
 
@@ -112,6 +119,77 @@ def kernel_smoothing(x, kernel, axis=-1):
         x = fftconvolve(x, kernel, **kw)
     return x
 
+
+def _acf(xd):
+    """Auto-correlation on a single time-series."""
+    n = len(xd)
+    acov = np.correlate(xd, xd, "full")[n - 1:] / n
+    return acov[: n + 1] / acov[0]
+
+
+def acf(x, axis=-1, demean=True):
+    """Compute the autocorrelation function.
+
+    Parameters
+    ----------
+    x : array_like
+        Array to use to estimate the auto-correlation. It can also be an
+        instance of xarray.DataArray.
+    axis : int, string | -1
+        Location of the temporal dimension. If the x input is a DataArray,
+        axis can be a string (e.g axis='times')
+    demean : bool | True
+        Specify whether the mean should be removed. Default is True.
+
+    Returns
+    -------
+    acf : array_like
+        Array of auto-correlation with the same shape and type as the input x
+    """
+    logger.info(f'Computing the auto-correlation function (demean={demean})')
+    # deal with xarray inputs
+    is_xarray = isinstance(x, xr.DataArray)
+    if is_xarray:
+        # get dimensions, coordinates and attributes
+        xdims, xcoords, xattrs = x.dims, x.coords, x.attrs
+        xattrs['demnean'] = int(demean)
+        # get axis number and name
+        if isinstance(axis, str):
+            axis_name = axis
+            axis = x.get_axis_num(axis)
+        elif isinstance(axis, int):
+            axis_name = x.dims[axis]
+        # drop labelling
+        x = x.data
+
+    # data demeaning
+    if demean:
+        xd = x - x.mean(axis=axis, keepdims=True)
+    else:
+        xd = x
+
+    # compute the auto-correlation
+    corr = np.apply_along_axis(_acf, axis, xd)
+
+    # xarray wrapping
+    if is_xarray:
+        # xarray transormation
+        corr = xr.DataArray(
+            corr, dims=xdims, coords=xcoords, attrs=xattrs, name='ACF'
+        )
+        # time axis should reflect auto-corr
+        times = corr[axis_name].data
+        dt = (times[1] - times[0])
+        corr[axis_name] = np.arange(len(times)) * dt
+
+    return corr
+
+
+###############################################################################
+###############################################################################
+#                                DATA SELECTION
+###############################################################################
+###############################################################################
 
 def nonsorted_unique(data, assert_unique=False):
     """Get an unsorted version of a element.
