### 2021-10-19 43fceb0
Implementation + test of the cross-correlation function
