+++ frites/conn/conn_ccf.py
@@ -1,0 +1,132 @@
+"""Cross-correlation function."""
+import numpy as np
+import xarray as xr
+
+from frites.conn import conn_io
+from frites.io import logger
+from frites.estimator import GCMIEstimator
+from frites.utils import parallel_func
+from frites.utils.preproc import _acf
+
+
+
+def conn_ccf(data, times=None, roi=None, normalized=True, n_jobs=1,
+             verbose=None):
+    """Single trial Cross-Correlation Function.
+
+    This function computes the pairwise Cross Correlation (CCF) at the single
+    trial level. This can be particulary usefull to find whether there are
+    temporal delays between times series.
+
+    Parameters
+    ----------
+    data : array_like
+        Electrophysiological data. Several input types are supported :
+
+            * Standard NumPy arrays of shape (n_epochs, n_roi, n_times)
+            * mne.Epochs
+            * xarray.DataArray of shape (n_epochs, n_roi, n_times)
+
+    times : array_like | None
+        Time vector array of shape (n_times,). If the input is an xarray, the
+        name of the time dimension can be provided
+    roi : array_like | None
+        ROI names of a single subject. If the input is an xarray, the
+        name of the ROI dimension can be provided
+    normalized : bool | True
+        Z-score normalization of the data. By default, it set to true.
+    n_jobs : int | 1
+        Number of jobs to use for parallel computing (use -1 to use all
+        jobs). The parallel loop is set at the pair level.
+
+    Returns
+    -------
+    ccf : array_like
+        The Cross-Correlation array of shape (n_epochs, n_pairs, n_times). When
+        the peak of correlation occurs at a negative time it means that the
+        target has to be moved **toward** the source. On the contrary, if the
+        peak occurs at positive time it means that the target is moved **away**
+        of the source.
+    """
+    # ________________________________ INPUTS _________________________________
+    # inputs conversion
+    data, cfg = conn_io(
+        data, times=times, roi=roi, agg_ch=False, win_sample=None, pairs=None,
+        sort=True, name='CCF', verbose=verbose,
+    )
+
+    # extract variables
+    x, trials, attrs = data.data, data['y'].data, cfg['attrs']
+    x_s, x_t = cfg['x_s'], cfg['x_t']
+    roi_p, roi_idx = cfg['roi_p'], cfg['roi_idx']
+    times = data['times'].data
+    n_pairs = len(x_s)
+
+    # data normalization
+    if normalized:
+        x = (x - x.mean(-1, keepdims=True)) / x.std(-1, keepdims=True)
+
+    # __________________________________ CCF __________________________________
+    # function to put in parallel
+    def para_fun(xs, xt):
+        n_trials = xs.shape[0]
+        corr = np.zeros((n_trials, int(2 * len(times)) - 1))
+        for n_t in range(n_trials):
+            corr[n_t, :] = _acf(xs[n_t, :], xt[n_t, :])
+        return corr
+
+    # prepare parallel function
+    n_jobs = 1 if n_pairs == 1 else n_jobs
+    parallel, p_fun = parallel_func(para_fun, n_jobs=n_jobs, verbose=verbose,
+                                    total=n_pairs, mesg='Estimating CCF')
+
+    logger.info(f'Computing CCF between {n_pairs} pairs')
+
+    # compute ccf
+    ccf = parallel(
+        p_fun(x[:, i_s, :], x[:, i_t, :]) for i_s, i_t in zip(x_s, x_t))
+    ccf = np.stack(ccf, axis=1)
+
[…]
