+++ docs/source/api/api_connectivity.rst
@@ -17,6 +17,7 @@ Connectivity metrics
 
    conn_dfc
    conn_covgc
+   conn_ccf
    conn_transfer_entropy
 
 Utility functions
