### 2021-10-19 bdbb59b
Improve temporal dimension of the ccf
