+++ frites/conn/conn_ccf.py
@@ -9,9 +9,17 @@ from frites.utils import parallel_func
 from frites.utils.preproc import _acf
 
 
+def para_fun(xs, xt):
+    """Single-trial ccf on a pair of signals."""
+    n_trials, n_times = xs.shape
+    corr = np.zeros((n_trials, int(2 * n_times) - 1))
+    for n_t in range(n_trials):
+        corr[n_t, :] = _acf(xs[n_t, :], xt[n_t, :])
+    return corr
+
 
 def conn_ccf(data, times=None, roi=None, normalized=True, n_jobs=1,
-             verbose=None):
+             times_as_sample=True, verbose=None):
     """Single trial Cross-Correlation Function.
 
     This function computes the pairwise Cross Correlation (CCF) at the single
@@ -35,6 +43,11 @@ def conn_ccf(data, times=None, roi=None, normalized=True, n_jobs=1,
         name of the ROI dimension can be provided
     normalized : bool | True
         Z-score normalization of the data. By default, it set to true.
+    times_as_sample : bool | False
+        Specify whether the time dimension of the cross-correlation output
+        should be described using the time unit of the input data or in
+        samples. By default, samples are used to describe lags between
+        time-series.
     n_jobs : int | 1
         Number of jobs to use for parallel computing (use -1 to use all
         jobs). The parallel loop is set at the pair level.
@@ -67,14 +80,6 @@ def conn_ccf(data, times=None, roi=None, normalized=True, n_jobs=1,
         x = (x - x.mean(-1, keepdims=True)) / x.std(-1, keepdims=True)
 
     # __________________________________ CCF __________________________________
-    # function to put in parallel
-    def para_fun(xs, xt):
-        n_trials = xs.shape[0]
-        corr = np.zeros((n_trials, int(2 * len(times)) - 1))
-        for n_t in range(n_trials):
-            corr[n_t, :] = _acf(xs[n_t, :], xt[n_t, :])
-        return corr
-
     # prepare parallel function
     n_jobs = 1 if n_pairs == 1 else n_jobs
     parallel, p_fun = parallel_func(para_fun, n_jobs=n_jobs, verbose=verbose,
@@ -89,13 +94,18 @@ def conn_ccf(data, times=None, roi=None, normalized=True, n_jobs=1,
 
     # ________________________________ OUTPUTS ________________________________
     # dataarray conversion
-    times_n = np.arange(ccf.shape[-1]).astype(float)# / cfg['sfreq']
+    times_n = np.arange(ccf.shape[-1]).astype(float)
     times_n -= times_n.mean()
+    unit = 'samples'
+    if not times_as_sample:
+        times_n /= cfg['sfreq']
+        unit = 'times'
     ccf = xr.DataArray(ccf, dims=('trials', 'roi', 'times'), name=f'CCF',
                        coords=(trials, roi_p, times_n))
 
     # add the windows used in the attributes
-    ccf.attrs = {**dict(type='ccf', normalized=int(normalized)), **attrs}
+    ccf.attrs = {**dict(type='ccf', normalized=int(normalized),
+                        times_unit=unit), **attrs}
 
     return ccf
 
@@ -123,7 +133,9 @@ if __name__ == '__main__':
     plt.figure(figsize=(15, 6))
 
     # compute delayed dfc
-    ccf = conn_ccf(x, times='times', roi='roi', n_jobs=1, verbose=False)
+    ccf = conn_ccf(x, times='times', roi='roi', n_jobs=1, verbose=False,
+                   times_as_sample=True)
+    print(ccf)
 
     plt.subplot(121)
     x.mean('trials').plot(x='times', hue='roi')
