### 2021-10-20 #3746 @brovelli
Thanks @madvn for accepting to review Frites. I hope you will find it interesting. In fact, I have seen you have developed an Information theory toolbox as well, also using PID measures. We are also currently working on them, probably adding them to the next release of Frites.
