### 2021-10-20 895a172
Add sources / targets indices to conn_* attributes
