### 2021-10-21 a2fdcf9
Include frequency checking in `conn_io`
