### 2021-10-21 9d54a4c
Smoothing parameters checked inside conn_io
