### 2021-10-21 cdb453f
Fix not using computed sampling frequency
