### 2021-10-21 de5d681
Reset to frequencies in samples
