### 2021-10-22 b990c76
Fix attribute conversion in conn related functions
