### 2021-10-22 #3842 @whedon
[REVIEW]: Frites: A Python package for functional connectivity analysis and group-level statistics of neurophysiological data
**Submitting author:** @EtienneCmb (<a href="http://orcid.org/0000-0002-7362-3247">Etienne Combrisson</a>)
**Repository:** https://github.com/brainets/frites
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.4.3

- [x] *Summary:** Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
