+++ requirements.txt
@@ -3,3 +3,5 @@ scipy
 mne
 joblib
 xarray
+netCDF4
+h5netcdf
