### 2021-11-02 #3842 @brovelli
We checked the 3 MISSING DOIs on https://dx.doi.org/ and they all appear to be correct.
