### 2021-11-05 75e5565
Adding neo to requirements and docs
