+++ requirements.txt
@@ -1,6 +1,7 @@
 numpy
 scipy
 mne
+neo
 joblib
 xarray
 netCDF4
