### 2021-11-05 760bbb2
move neo handling from ds_ephy to suj_ephy
