+++ frites/dataset/__init__.py
@@ -2,7 +2,7 @@
 
 This submodule includes containers for the neurophysiological data either for
 a single-subject or multiple subjects. Several input types are supported
-(NumPy, MNE, Xarray).
+(NumPy, MNE, Neo, Xarray).
 """
 from .suj_ephy import SubjectEphy  # noqa
 from .ds_ephy import DatasetEphy  # noqa
