+++ frites/simulations/tests/test_sim_generate_data.py
@@ -26,9 +26,8 @@ class TestGenerateData(object):
         data, _, _ = sim_single_suj_ephy(as_mne=True)
         assert isinstance(data, EpochsArray)
         # neo type
-        if HAS_NEO:
-            data, _, _ = sim_single_suj_ephy(as_neo=True)
-            assert isinstance(data, neo.Block)
+        data, _, _ = sim_single_suj_ephy(as_neo=True)
+        assert isinstance(data, neo.Block)
 
     def test_sim_multi_suj_ephy(self):
         """Test function sim_multi_suj_ephy."""
@@ -45,7 +44,6 @@ class TestGenerateData(object):
         data, _, _ = sim_multi_suj_ephy(n_subjects=5, as_mne=True)
         assert all([isinstance(k, EpochsArray) for k in data])
         # neo type
-        if HAS_NEO:
-            data, _, _ = sim_multi_suj_ephy(n_subjects=5, as_neo=True)
-            assert all([isinstance(k, neo.Block) for k in data])
+        data, _, _ = sim_multi_suj_ephy(n_subjects=5, as_neo=True)
+        assert all([isinstance(k, neo.Block) for k in data])
 
