### 2021-11-05 2248ce2
starting to add neo support to conn
