### 2021-11-05 a21de23
adding 'as_neo' support for single subject simulation
