### 2021-11-05 cf104d7
listing neo as possible input type
