### 2021-11-05 d28983a
make neo a soft dependency
