+++ frites/dataset/tests/test_suj_ephy.py
@@ -1,10 +1,15 @@
 """Test SubjectEphy and internal conversions."""
+import pytest
 import numpy as np
 import xarray as xr
 import pandas as pd
 import mne
-import neo
-import quantities as pq
+try:
+    import neo
+    import quantities as pq
+    HAVE_NEO = True
+except ModuleNotFoundError:
+    HAVE_NEO = False
 
 from frites.dataset import SubjectEphy
 from frites.utils.perf import id as id_arr
@@ -49,7 +54,8 @@ class TestSubjectEphy(object):  # noqa
         elif (dtype == 'mne') and (ndim == 4):
             info = mne.create_info(ch_names, sfreq, ch_types='seeg')
             x_out = mne.time_frequency.EpochsTFR(info, x_4d, times, freqs)
-        elif (dtype == 'neo'):
+        elif dtype == 'neo':
+            assert HAVE_NEO, 'Requires Neo to be installed'
             data = x_3d if ndim == 3 else x_4d
             block = neo.Block()
             for epoch_id in range(len(x_3d)):
@@ -130,6 +136,7 @@ class TestSubjectEphy(object):  # noqa
         da_4d = SubjectEphy(mne_4d, y=y_int, z=z, roi=roi, times=times, **kw)
         self._test_memory(x_4d, da_4d.data)
 
+    @pytest.mark.skipif(not HAVE_NEO, reason="requires Neo")
     def test_neo_inputs(self):
         """Test function neo_inputs."""
         # ___________________________ test 3d inputs __________________________
