+++ frites/simulations/tests/test_sim_generate_data.py
@@ -4,9 +4,9 @@ from mne import EpochsArray
 
 try:
     import neo
-    HAS_NEO = True
+    HAVE_NEO = True
 except ModuleNotFoundError:
-    HAS_NEO = False
+    HAVE_NEO = False
 
 from frites.simulations import (sim_single_suj_ephy, sim_multi_suj_ephy)
 
@@ -26,8 +26,9 @@ class TestGenerateData(object):
         data, _, _ = sim_single_suj_ephy(as_mne=True)
         assert isinstance(data, EpochsArray)
         # neo type
-        data, _, _ = sim_single_suj_ephy(as_neo=True)
-        assert isinstance(data, neo.core.Block)
+        if HAVE_NEO:
+            data, _, _ = sim_single_suj_ephy(as_neo=True)
+            assert isinstance(data, neo.core.Block)
 
     def test_sim_multi_suj_ephy(self):
         """Test function sim_multi_suj_ephy."""
@@ -44,6 +45,7 @@ class TestGenerateData(object):
         data, _, _ = sim_multi_suj_ephy(n_subjects=5, as_mne=True)
         assert all([isinstance(k, EpochsArray) for k in data])
         # neo type
-        data, _, _ = sim_multi_suj_ephy(n_subjects=5, as_neo=True)
-        assert all([isinstance(k, neo.Block) for k in data])
+        if HAVE_NEO:
+            data, _, _ = sim_multi_suj_ephy(n_subjects=5, as_neo=True)
+            assert all([isinstance(k, neo.Block) for k in data])
 
