+++ requirements.txt
@@ -1,7 +1,6 @@
 numpy
 scipy
 mne
-neo
 joblib
 xarray
 netCDF4
