### 2021-11-15 11bc06d
Improve WfMi message about copnorm
