### 2021-11-27 #3842 @travisbthomp
@EtienneCmb  I am running some of the examples from the documentation.  I noticed the following (likely unintentional) output from plot_stim_spec_network.py

plot_stim_spec_network.py:211: SyntaxWarning: "is" with a literal. Did you mean "=="?
  if direction is 'x->y':
plot_stim_spec_network.py:213: SyntaxWarning: "is" with a literal. Did you mean "=="?
  elif direction is 'y->x':
plot_stim_spec_network.py:215: SyntaxWarning: "is" with a literal. Did you mean "=="?
  elif direction is 'x.y':
plot_stim_spec_network.py:260: SyntaxWarning: "is" with a literal. Did you mean "=="?
  if direction is 'x->y':
plot_stim_spec_network.py:262: SyntaxWarning: "is" with a literal. Did you mean "=="?
  elif direction is 'y->x':
plot_stim_spec_network.py:264: SyntaxWarning: "is" with a literal. Did you mean "=="?
  elif direction is 'x.y':
