### 2021-11-27 #3842 @travisbthomp
@EtienneCmb  The second figure generated  by plot_wf_mi_cc.py does not match the second figure in the corresponding documentation.
