### 2021-11-28 #3842 @madvn
Hello all, I am just getting to this as well.

Looks like a very interesting toolbox. I think the idea of feature-specific functional networks using mutual information can be used widely in not just neuroscientific settings but also sociology or even, more topically, epidemiology. I've updated my checklist based on a quick look at the repo and installing the package. I also took a quick look at the paper and have some notes for the authors based on the following requirements

You have a summary but please revise it to address the "for a diverse, non-specialist audience" part of the requirement. Add a note about the breadth of usability of these tools in other domains besides neuroscience.

Please place your tool in the space of other information theory tools such as `infotheory`, `IDTXL` etc. as well as network neuroscience tools such as Brain connectivity toolbox etc. I do believe there are unique aspects to your package, such as combining information-theoretic measures with network neuroscience to find feature-specific functional networks (I'm not aware of other tools that does this -- I will look it up. If you are aware of others, please include those.

Please update references based on additional paper that you will cite to address the above points.
