### 2021-11-29 6fa71bf
Replace 'is' by '==' (Python >3.6)
