+++ frites/__init__.py
@@ -9,7 +9,7 @@ import logging
 from frites import (io, core, conn, stats, utils, workflow, simulations,  # noqa
                     estimator)
 
-__version__ = "0.4.0"
+__version__ = "0.4.1"
 
 # -----------------------------------------------------------------------------
 # Set 'info' as the default logging level
