### 2021-11-29 51bc90c
Use makefile to generate the dist
