+++ .github/workflows/python-publish.yml
@@ -27,5 +27,7 @@ jobs:
         TWINE_USERNAME: ${{ secrets.PYPI_USERNAME }}
         TWINE_PASSWORD: ${{ secrets.PYPI_PASSWORD }}
       run: |
-        python setup.py sdist bdist_wheel
-        twine upload dist/*
+        make clean_dist
+        make build_dist
+        make check_dist
+        make upload_dist
