### 2021-11-29 #3842 @EtienneCmb
Dear @travisbthomp,

> plot_stim_spec_network.py:211: SyntaxWarning: "is" with a literal. Did you mean "=="?

I fixed this morning the warning message by replacing 'is' with '=='

> The second figure generated by plot_wf_mi_cc.py does not match the second figure in the corresponding documentation.

Indeed, as you saw, we are mainly using random data for illustrating the functions

> The example plot_est_custom.py does not work for me and generates the following error
> The example plot_ccf.py does not work for me and generates the following error

Indeed, both examples were using the develop branch. This morning I pushed the version 0.4.1 of Frites to Pypi, it should works now once updated.
