+++ .github/workflows/test_doc.yml
@@ -54,7 +54,7 @@ jobs:
           touch build/html/.nojekyll
 
       - name: Deploy Github Pages 🚀
-        uses: JamesIves/github-pages-deploy-action@4.1.6
+        uses: JamesIves/github-pages-deploy-action@4.1.7
         with:
           branch: gh-pages
           folder: docs/build/html/
