### 2021-11-30 ba44a42
Integration of the function for simulated spatio-temporal ground-truths
