+++ examples/simulations/plot_ground_truth.py
@@ -1,0 +1,260 @@
+"""
+Generate spatio-temporal ground-truths
+======================================
+
+Frites provides some functions to generate spatio-temporal ground-truths (i.e.
+effects that are distributed across time and space with a predefined profile).
+Those ground-truths can be particularly interesting to simulate the data coming
+from multiple subjects, compare group-level strategies such as methods for
+multiple comparisons.
+
+This example illustrates :
+
+    * How the predefined ground-truth effects looks like
+    * How to generate the data coming from multiple subjects
+    * How to use the workflow of mutual-information to retrieve the effect
+    * How to use statistical measures to compare the effect detected as
+      significant and the ground-truth
+"""
+import numpy as np
+import xarray as xr
+
+from frites.simulations import sim_ground_truth
+from frites.dataset import DatasetEphy
+from frites.workflow import WfMi
+
+import matplotlib.pyplot as plt
+import matplotlib as mpl
+
+plt.style.use('ggplot')
+
+
+###############################################################################
+# Comparison of the implemented ground truth
+# ------------------------------------------
+#
+# In this first part we illustrate the implemented ground-truths profiles. This
+# includes effects with varying covariance over time and space, weak and
+# diffuse effects and strong and focal effect.
+
+gtypes = ['tri', 'tri_r', 'diffuse', 'focal']
+n_subjects = 1
+n_epochs = 5
+
+gts = {}
+for gtype in gtypes:
+    gts[gtype] = sim_ground_truth(n_subjects, n_epochs, gtype=gtype,
+                                  gt_only=True, gt_as_cov=True)
+
+# plot the four ground-truths
+fig, axs = plt.subplots(
+    ncols=len(gtypes), sharex=False, sharey=False, figsize=(18, 4.5),
+    gridspec_kw=dict(wspace=.2, left=0.05, right=0.95)
+)
+axs = np.ravel(axs)
+
+for n_g, g in enumerate(gtypes):
+    plt.sca(axs[n_g])
+    df = gts[g].to_pandas().T
+    plt.pcolormesh(df.columns, df.index, df.values, cmap='Spectral_r',
+                   shading='nearest', vmin=0., vmax=0.3)
+    plt.title(g.capitalize(), fontweight='bold', fontsize=15)
+    plt.grid(True)
+    plt.xlabel('Time (bin)')
+    if n_g == 0: plt.ylabel('Spatial (bin)')
+    plt.colorbar()
+
+plt.show()
+
+###############################################################################
+# .. note::
+#       Here, the ground-truth contains values of covariance at specific time
+#       and spatial bins. The covariance reflects where there's a relation
+#       between the brain data and the external continuous y variable and how
+#       strong is this relation.
+
+
+###############################################################################
+# Data simulation
+# ---------------
+#
+# In this second part, we use the same function to simulate the data coming
+# from multiple subjects. The returned data are a list of length n_subjects
+# where each element of the list is an array (xarray.DataArray) of shape
+# (n_epochs, n_roi, n_times). In addition to the simulated data, the external
+# continuous variable is set as a coordinate of the trial dimension ('y')
+
+gtype = 'tri'    # ground truth type
+n_subjects = 10  # number of simulated subjects
[…]
