+++ frites/simulations/sim_local_mi.py
@@ -1,5 +1,8 @@
 """Simulate local representation of mutual information."""
 import numpy as np
+import xarray as xr
+
+from frites.io import logger, set_log_level
 
 
 """
@@ -386,3 +389,167 @@ def sim_local_ccd_ss(n_epochs=10, n_times=100, n_roi=1, n_conditions=2,
         cl_cov=cl_cov, random_state=random_state)
     z = np.random.randint(0, n_conditions, (n_epochs,))
     return x, y, z, roi, times
+
+
+def sim_ground_truth(n_subjects, n_epochs, gtype='tri', perc=100, p_pos=1.,
+                     gt_as_cov=False, gt_only=False, random_state=None,
+                     verbose=None):
+    """Spatio-temporal ground truth simulation.
+
+    This function can be used to simulate the data coming from multiple
+    subjects with specific ground-truth arrays with clusters distributed over
+    time ad space. This function typically returned two outputs :
+
+        * The simulated data (i.e. the brain data and a continuous external
+          variable)
+        * The ground-truth array which specify where the effects are and the
+          strength of the effects
+
+    The outputs of this function can be used to compare different statistical
+    models and methods for correcting for multiple comparisons.
+
+    Parameters
+    ----------
+    n_subjects, n_epochs : int
+        Number of subjects and trials
+    gtype : {'tri', 'tri_r', 'diffuse', 'focal'}
+        Ground-truth type. Choose either :
+
+            * 'tri' or 'tri_r' : spatially diffuse effects
+            * 'diffuse' : weak effect spatially diffuse
+            * 'focal' : strong and spatially focalized effect
+    perc : int | 100
+        Percentage of subjects having the effect (equivalent to inter-subject
+        consistency). For example, if n_subjects=10 and perc is 80%, 8 subjects
+        out of the 10 will have the effect. This parameters can be useful to
+        investigate how many subjects are necessary to retrieve the effect.
+    p_pos : float | 1.
+        Proportion of positive correlation. A value of 1 indicates that all of
+        the subjects have positive correlations between the simulated brain
+        data and the regressor. Conversely, a value of 0. indicates that all
+        of the subjects have negative correlations. In between, for example
+        with a value 0.5 indicates that half of the subjects will have positive
+        correlations and the other half, negative correlations.
+    gt_as_cov : bool | False
+        Specify whether the returned ground-truth array should contained
+        boolean values indicating whether there is an effect or not at a
+        specific time-space bin (False). If True, the ground-truth array will
+        contained the values used for covariance.
+    gt_only : bool | False
+        Specify whether only the ground-truth should be returned
+    random_state : int | None
+        Fix the random state of the machine. This parameter can be useful for
+        having reproducible results. If None, the seed is randomly picked.
+
+    Returns
+    -------
+    da : list
+        List containing the simulated data coming from multiple subjects. Each
+        element of the list represents a single subject of shape
+        (n_epoch, n_roi, n_times). This argument is returned if gt_only is set
+        to False.
+    gt : array_like
+        Ground-truth array of shape (n_roi, n_times)
+    """
+    set_log_level(verbose)
+    assert gtype in ['tri', 'tri_r', 'diffuse', 'focal']
+    assert 0 <= perc <= 100
[…]
