### 2021-12-01 #3842 @travisbthomp
# Suggestions for paper

After reviewing the paper I would like to make a few comments in the spirit of the review guidelines above.

### Summary
#### Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

The description of the high level functionality is not quite clear.  "Frites provides such an integrated framework" for what? Is Frites limited to "discovery of cognitive brain networks from multi-channel neurophysiological datasets"? What types of networks?  Conversely, I think that the documentation text is quite approachable and nice!

> Frites is a Python toolbox for assessing information-based measures on human and animal
> neurophysiological data (M/EEG, Intracranial). The toolbox also includes directed and undirected
> connectivity metrics such as group-level statistics on measures of information (information-theory,
> machine-learning and measures of distance).

In short, the current summary seems obfuscated and potentially difficult to read for a "non-specialist audience".

### Statement of need
#### Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?

The paper does indeed have a section titled 'Statement of Need'.  However, this section does not establish a clear need for the software nor does it "clearly state what problems the software is designed to solve and who the target audience is".  At least, it does not do these things in a way that can be appreciated by an non-expert in the area, such as myself, who is not currently aware of the state of the art approaches, what problems they could solve better, and why Frites fills any of these gaps.

I would suggest revisiting this write up with the journal guidelines in mind.

### State of the field
#### Do the authors describe how this software compares to other commonly-used packages?

I cannot find where this has been done in a manner that is decipherable to a non-expert.  I would suggest an approach which clearly states the central problem Frites aims to solve and endeavors a comparative discussion to other available tools; what are the advantages of Frites over other packages?  If there are no other packages, simply state that Frites is the first of its kind.

### Quality of writing
#### Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

The authors do not struggle with the tenets of structural composition.  The central message of the statement of need is not currently clear, but this is not an issue with the sentence-by-sentence structure/language/quality per se.

### References
#### Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?

I did not see any issues with the references and all provided DOI links correct redirected to the appropriate paper.
