### 2021-12-01 #3842 @EtienneCmb
Thank you @madvn and @travisbthomp for the comments on the manuscript, I do agree with both of you about the lack of accessibility. I'll work on the manuscript soon to clarify the text and make it as accessible as possible to the non-expert.
