### 2021-12-01 #3842 @travisbthomp
## Functionality Review

### Functionality
#### Have the functional claims of the software been confirmed?

It is not currently clear what primary claims should be assessed.  I was able to piece together a possibly-incomplete list from the documentation and accompanying paper.

From the documentation I can see a few things:

1. Frites is a Python toolbox for assessing information-based measures on human and animal neurophysiological data (M/EEG, Intracranial)
2. Extract task-related cognitive brain networks : i.e brain regions and connectivity between brain regions that are modulated according to the task
3. Group-level statistical inferences : being able to draw conclusions on a population of subjects while taking into account the the richness of data at the single subject level
4. With Frites, we try to isolate the brain activity that is either related to the stimulus or to the behavior.

From the paper, there are a few more claims to be assessed:

1. this package can be used to study the relation between local brain activity, cross-frequency coupling and inter-areal functional connectivity (FC) with experimental variables
2.  For what concerns FC, the toolbox allows the estimate of
  a. dynamic (i.e., time-resolve) functional connectivity
  b. undirected (e.g., mutual information) functional connectivity and
  c. directed (e.g. Granger causality) functional connectivity on a single-trial basis
3. The package integrates a non-parametric permutation-based statistical framework to perform group-level inferences on non negative measures of information.
4. The toolbox includes different methods that cope with multiple-comparison correction problems such as test- and cluster-wise p-value corrections.
5. The implemented framework supports both fixed- and random-effect models to adapt to inter-individuals and inter-sessions variability

The authors should review this list to make sure that I have not misrepresented, or missed any of, their functionality claims.  It seems that verifying these functionality claims would be easiest by pointing to the corresponding example code.  To make these claims easier to verify I would suggest considering a strategy similar to the following:

On the [main documentation page](https://brainets.github.io/frites/index.html) it might be natural to have a section called something like "What Frites can do for you" between "Description" and "Highlights" which lists the functionality claims and direct link to a corresponding script from the examples section.  For instance

"What Frites can do for you"
1. Frites can assess information-based measures on human and animal neurophysiological data (as a link to an example script)
2. Frites can extract task-related cognitive brain networks (as a link to an example script)
3. Frites can conduct a group-level statistical inference on time-series data (as a link to an example script)
4. Frites can be used to study relationships between local brain activity and cross-frequency coupling (link to example script)
5. Frites can also highlight links between inter-areal functional connectivity  with experimental variables (link to example script)
and so forth.  Essentially, each claim can be made explicit and each claim can be linked to a demonstration of that claim vis-a-vis one of the example scripts.

### Performance
#### If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

There are a few performance-related mentions that I can find in the paper.  They are as follows (performance-related words are in **bold** text)

> Frites is **optimally designed** for the analysis of continuous and multi-channel
> neurophysiological data, encompassing recordings with either uniform spatial
> sampling (e.g., M/EEG data) and spatially sparse intracranial recordings, such
> as intracranial EEG or Local Field Potentials (LFPs).

The claim here appears to be that the design of Frites is *optimal* when considering the task of analyzing particular types of data sets.  What does optimal mean here and where can evidence be found for this claim?  I did not see, for instance, a documentation example where Frites is compared to other software to show that it is "optimal" amongst the current choices.

Maybe this claim could be made more explicit and confirmed as part of the documentation via a "Performance" page that could be linked off the main bar at the top of the documentation (which currently holds "Overview" "Installation" "API Reference" "Examples")

> In order to facilitate **automated** and **efficient usage**, Frites provides a set of
> high-level workflows that integrate several analysis steps from information-based
> estimation to network level statistical inference.

Similar to above.  Where do I find any real information about what workflows really are, let alone how they are *automated* and enable *more efficiency*? I assume that would be [here](https://brainets.github.io/frites/overview/ovw_frites.html#main-frites-workflows) but this page only has a cursory discussion of workflows and leaves me with no real concept of what they actually are beyond whatever my intuition suggests from the name "workflow".  There is an "in a nutshell" discussion which is not particuarly helpful in addition to a link to "chekout those examples" which takes me to the main examples page so that I must dig to find examples that demonstrate the workflows themselves.  Moreover, the links to frites.dataset, frites.workflow and frites.conn on the [overview page](https://brainets.github.io/frites/overview/ovw_frites.html)  (above the workflow section) all end in a 404.

The burden of figuring out what a workflow is, and substantiating the claims of automation and efficiency should not be on the reader.

> Since several computations implemented in the workflows, such as permutation tests,
> are **computationally demanding**, Frites natively supports parallel processing using the
> Joblib package.

Which computations are the most demanding? Can we point to the examples directly in addition to any performance gains from using Joblib?  This would fit well on a "Performance" documentation page as mentioned above

> In addition, some functions can further be **accelerated** using the Numba compiler
> as an optional dependency. Programming **optimizations** and external dependencies allow to
> investigate large-scale datasets in a **reasonable** time.

What functions can be accelerated / where is acceleration most profitable in Frites?  To justify the claim of "large scale datasets in reasonable time" it seems natural to include a benchmark plot showing the processing time of Frites, versus other software packages as well, against a data set size.
