### 2021-12-03 46a18b1
Update layout of the example
