### 2021-12-08 5278eb4
Add files via upload
Adding the spectral connectivity methods to frites.conn and the respective unit tests.
