+++ tests/test_conn_spec.py
@@ -1,0 +1,176 @@
+import numpy as np
+from frites.conn.conn_tf import _tf_decomp
+from frites.conn.conn_spec import conn_spec
+
+
+class TestConnSpec:
+
+    n_roi, n_times, n_epochs = 4, 1000, 20
+    n_edges = int(n_roi*(n_roi-1)/2)
+    sfreq, freqs = 200, np.arange(1, 51, 1)
+    n_freqs = len(freqs)
+    n_cycles = freqs/2
+    times = np.arange(0, n_times//sfreq, 1/sfreq)
+    eta = np.random.normal(0, 1, size=(n_epochs, n_roi, n_times))
+
+    def test_tf_decomp(self, ):
+
+        # Test output shape
+        for mode in ["morlet", "multitaper"]:
+            out = _tf_decomp(self.eta, self.sfreq, self.freqs, mode=mode,
+                             n_cycles=self.n_cycles, n_jobs=1)
+            self.__assert_shape(out.shape, conn=False)
+
+        # For multitaper test both single and array mt_bandwidth
+        out1 = _tf_decomp(self.eta, self.sfreq, self.freqs, mode="multitaper",
+                          n_cycles=self.n_cycles, mt_bandwidth=4, n_jobs=1)
+        out2 = _tf_decomp(self.eta, self.sfreq, self.freqs, mode="multitaper",
+                          n_cycles=self.n_cycles,
+                          mt_bandwidth=[4]*self.n_freqs, n_jobs=1)
+        np.testing.assert_array_equal(out1, out2)
+        ##################################################################
+        # Compare the auto-spectra with groundtruth
+        ##################################################################
+        for mode in ["morlet", "multitaper"]:
+
+            # 1. Compare for stationary sinal
+            x = self.__get_signal(stationary=True)
+
+            out = _tf_decomp(x, self.sfreq, self.freqs, mode=mode,
+                             n_cycles=self.n_cycles, n_jobs=1)
+            out = (out*np.conj(out)).real
+
+            if mode == "morlet":
+                val, atol = 20, 2
+            else:
+                val, atol = 5.8, 0.3
+            actual = out.mean(
+                axis=(0, -1))[:, self.__get_freqs_indexes(28, 32)].mean(1)
+            np.testing.assert_allclose(
+                actual, val*np.ones_like(actual), atol=atol)
+
+            # 2. Compare for non-stationary signal
+            x = self.__get_signal(stationary=False)
+
+            out = _tf_decomp(x, self.sfreq, self.freqs, mode=mode,
+                             n_cycles=self.n_cycles, n_jobs=1)
+            out = (out*np.conj(out)).real
+
+            if mode == "morlet":
+                val, atol = 11, 1
+            else:
+                val, atol = 3.2, 0.3
+            actual1 = out.mean(
+                axis=(0, -1))[:, self.__get_freqs_indexes(8, 12)].mean(1)
+            actual2 = out.mean(
+                axis=(0, -1))[:, self.__get_freqs_indexes(28, 32)].mean(1)
+            np.testing.assert_allclose(
+                actual1, val*np.ones_like(actual), atol=atol)
+            np.testing.assert_allclose(
+                actual2, val*np.ones_like(actual), atol=atol)
+
+    def test_conn_spec(self,):
+        """Test function conn_spec"""
+
+        # General parameters for the conn_spec function
+        kw = dict(sfreq=self.sfreq, freqs=self.freqs, n_jobs=1, verbose=False,
+                  n_cycles=self.n_cycles, times=self.times, sm_kernel='square')
+
+        for method in ['coh', 'plv']:
+            ##################################################################
+            # Check general attributes of the conn_spec container
+            ##################################################################
[…]
