### 2021-12-08 1048d69
Revert "Adding spectral connectivity metric to frites.conn."
