+++ conn_spec.py
@@ -1,309 +1,0 @@
-"""Compute single-trial spectral connectivity.
-
-Authors : Vinicius Lima <vinicius.lima.cordeiro@gmail.com >
-          Etienne Combrisson <e.combrisson@gmail.com>
-
-License : BSD (3-clause)
-"""
-import numpy as np
-import xarray as xr
-
-from frites.conn import conn_io
-from frites.io import set_log_level, logger, check_attrs
-from frites.utils import parallel_func
-from .conn_tf import (_tf_decomp, _create_kernel,
-                      _smooth_spectra, _foi_average)
-
-
-###############################################################################
-###############################################################################
-#                               CORE FUNCTIONS
-###############################################################################
-###############################################################################
-
-def _coh(w, kernel, foi_idx, x_s, x_t, kw_para):
-    """Pairwise coherence."""
-    # auto spectra (faster that w * w.conj())
-    s_auto = w.real ** 2 + w.imag ** 2
-
-    # smooth the auto spectra
-    s_auto = _smooth_spectra(s_auto, kernel)
-
-    # define the pairwise coherence
-    def pairwise_coh(w_x, w_y):
-        # computes the coherence
-        s_xy = w[:, w_y, :, :] * np.conj(w[:, w_x, :, :])
-        s_xy = _smooth_spectra(s_xy, kernel)
-        s_xx = s_auto[:, w_x, :, :]
-        s_yy = s_auto[:, w_y, :, :]
-        out = np.abs(s_xy) ** 2 / (s_xx * s_yy)
-        # mean inside frequency sliding window (if needed)
-        if isinstance(foi_idx, np.ndarray):
-            return _foi_average(out, foi_idx)
-        else:
-            return out
-
-    # define the function to compute in parallel
-    parallel, p_fun = parallel_func(pairwise_coh, **kw_para)
-
-    # compute the single trial coherence
-    return parallel(p_fun(s, t) for s, t in zip(x_s, x_t))
-
-
-def _plv(w, kernel, foi_idx, x_s, x_t, kw_para):
-    """Pairwise phase-locking value."""
-    # define the pairwise plv
-    def pairwise_plv(w_x, w_y):
-        # computes the plv
-        s_xy = w[:, w_y, :, :] * np.conj(w[:, w_x, :, :])
-        # complex exponential of phase differences
-        exp_dphi = s_xy / np.abs(s_xy)
-        # smooth e^(-i*\delta\phi)
-        exp_dphi = _smooth_spectra(exp_dphi, kernel)
-        # computes plv
-        out = np.abs(exp_dphi)
-        # mean inside frequency sliding window (if needed)
-        if isinstance(foi_idx, np.ndarray):
-            return _foi_average(out, foi_idx)
-        else:
-            return out
-
-    # define the function to compute in parallel
-    parallel, p_fun = parallel_func(pairwise_plv, **kw_para)
-
-    # compute the single trial coherence
-    return parallel(p_fun(s, t) for s, t in zip(x_s, x_t))
-
-
-def _cs(w, kernel, foi_idx, x_s, x_t, kw_para):
-    """Pairwise cross-spectra."""
-    # define the pairwise cross-spectra
-    def pairwise_cs(w_x, w_y):
-        #  computes the cross-spectra
-        out = w[:, w_x, :, :] * np.conj(w[:, w_y, :, :])
-        out = _smooth_spectra(out, kernel)
-        if foi_idx is not None:
-            return _foi_average(out, foi_idx)
-        else:
-            return out
-
-    # define the function to compute in parallel
-    parallel, p_fun = parallel_func(pairwise_cs, **kw_para)
-
-    # compute the single trial coherence
[…]
