### 2021-12-08 084d9c1
Adding spectral conn methods
