+++ frites/conn/conn_tf.py
@@ -1,0 +1,217 @@
+"""Compute time-frequency decomposition base on Morlet or Multitaper methods.
+
+This script contains the function:
+
+1. _tf_decomp used to decompose the sinal in tf domains using Morlet or
+    Multitaper
+1. _create_kernel: Create a kernel to smooth the spectra (either boxcar or
+    hanning)
+2. _smooth_kernel: Perform the smoothing operation on the spectra based on the
+   convolution theorem
+"""
+# Authors : Vinicius Lima <vinicius.lima.cordeiro@gmail.com >
+#           Etienne Combrisson <e.combrisson@gmail.com>
+#
+# License : BSD (3-clause)
+
+import xarray as xr
+import numpy as np
+
+from mne.time_frequency import tfr_array_morlet, tfr_array_multitaper
+from scipy.signal import fftconvolve
+
+from frites.io import logger
+
+
+def _tf_decomp(data, sf, freqs, mode='morlet', n_cycles=7.0, mt_bandwidth=None,
+               decim=1, kw_cwt={}, kw_mt={}, n_jobs=1):
+    """Time-frequency decomposition using MNE-Python.
+    Parameters
+    ----------
+    data : array_like
+        Electrophysiological data of shape (n_trials, n_chans, n_times)
+    sf : float
+        Sampling frequency
+    freqs : array_like
+        Central frequency vector.
+    mode : {'morlet', 'multitaper'}
+        Spectrum estimation mode can be either: 'multitaper' or 'morlet'.
+    n_cycles : array_like | 7.
+        Number of cycles to use for each frequency. If a float or an integer is
+        used, the same number of cycles is going to be used for all frequencies
+    mt_bandwidth : int | float | array_like | None
+        The bandwidth of the multitaper windowing function in Hz. Only used in
+        'multitaper' mode.
+    decim : int | 1
+        To reduce memory usage, decimation factor after time-frequency
+        decomposition. default 1 If int, returns tfr[…, ::decim]. If slice,
+        returns tfr[…, decim].
+    kw_cwt : dict | {}
+        Additional arguments sent to the mne-function
+        :py:`mne.time_frequency.tfr_array_morlet`
+    kw_mt : dict | {}
+        Additional arguments sent to the mne-function
+        :py:`mne.time_frequency.tfr_array_multitaper`
+    Returns
+    -------
+    out : array_like
+        Time-frequency transform of shape (n_epochs, n_chans, n_freqs, n_times)
+    """
+    if mode == 'morlet':
+        out = tfr_array_morlet(
+            data, sf, freqs, n_cycles=n_cycles, output='complex', decim=decim,
+            n_jobs=n_jobs, **kw_cwt)
+    elif mode == 'multitaper':
+        # In case multiple values are provided for mt_bandwidth
+        # the MT decomposition is done separatedly for each
+        # Frequency center
+        if isinstance(mt_bandwidth, (list, tuple, np.ndarray)):
+            # Arrays freqs, n_cycles, mt_bandwidth should have the same size
+            assert len(freqs) == len(n_cycles) == len(mt_bandwidth)
+            out = []
+            for f_c, n_c, mt in zip(freqs, n_cycles, mt_bandwidth):
+                out += [tfr_array_multitaper(
+                    data, sf, [f_c], n_cycles=float(n_c), time_bandwidth=mt,
+                    output='complex', decim=decim, n_jobs=n_jobs, **kw_mt)]
+            out = np.stack(out, axis=2).squeeze()
+        elif isinstance(mt_bandwidth, (type(None), int, float)):
+            out = tfr_array_multitaper(
+                data, sf, freqs, n_cycles=n_cycles,
+                time_bandwidth=mt_bandwidth, output='complex', decim=decim,
+                n_jobs=n_jobs, **kw_mt)
+    else:
[…]
