+++ frites/conn/tests/test_conn_spec.py
@@ -1,8 +1,7 @@
 import numpy as np
-import xarray as xr
-import matplotlib.pyplot as plt
-from .conn.conn_tf import _tf_decomp
-from .conn.conn_spec import conn_spec
+
+from frites.conn.conn_tf import _tf_decomp
+from frites.conn.conn_spec import conn_spec
 
 
 class TestConnSpec:
@@ -82,8 +81,8 @@ class TestConnSpec:
             # Test container attributes, dims and coords
             assert out.name == method
             self.__assert_shape(out.shape)
-            self.__assert_default_rois(out.roi.data)
-            self.__assert_dims(out.dims)
+            self.__asseis(out.roi.data)
+            self.__assert_dims(rt_default_roout.dims)
             self.__assert_attrs(out.attrs)
             
             ##################################################################
