### 2021-12-08 2084158
Fix assert default rois
