### 2021-12-08 2aad81e
Fix input to assert_dims
