+++ frites/conn/tests/test_conn_spec.py
@@ -7,11 +7,11 @@ from frites.conn.conn_spec import conn_spec
 class TestConnSpec:
 
     n_roi, n_times, n_epochs = 4, 1000, 20
-    n_edges = int(n_roi*(n_roi-1)/2)
+    n_edges = int(n_roi * (n_roi - 1) / 2)
     sfreq, freqs = 200, np.arange(1, 51, 1)
     n_freqs = len(freqs)
-    n_cycles = freqs/2
-    times = np.arange(0, n_times//sfreq, 1/sfreq)
+    n_cycles = freqs / 2
+    times = np.arange(0, n_times // sfreq, 1 / sfreq)
     eta = np.random.normal(0, 1, size=(n_epochs, n_roi, n_times))
 
     def test_tf_decomp(self, ):
@@ -27,7 +27,7 @@ class TestConnSpec:
                           n_cycles=self.n_cycles, mt_bandwidth=4, n_jobs=1)
         out2 = _tf_decomp(self.eta, self.sfreq, self.freqs, mode="multitaper",
                           n_cycles=self.n_cycles,
-                          mt_bandwidth=[4]*self.n_freqs, n_jobs=1)
+                          mt_bandwidth=[4] * self.n_freqs, n_jobs=1)
         np.testing.assert_array_equal(out1, out2)
         ##################################################################
         # Compare the auto-spectra with groundtruth
@@ -39,7 +39,7 @@ class TestConnSpec:
 
             out = _tf_decomp(x, self.sfreq, self.freqs, mode=mode,
                              n_cycles=self.n_cycles, n_jobs=1)
-            out = (out*np.conj(out)).real
+            out = (out * np.conj(out)).real
 
             if mode == "morlet":
                 val, atol = 20, 2
@@ -55,7 +55,7 @@ class TestConnSpec:
 
             out = _tf_decomp(x, self.sfreq, self.freqs, mode=mode,
                              n_cycles=self.n_cycles, n_jobs=1)
-            out = (out*np.conj(out)).real
+            out = (out * np.conj(out)).real
 
             if mode == "morlet":
                 val, atol = 11, 1
@@ -65,9 +65,9 @@ class TestConnSpec:
                 axis=(0, -1))[:, self.__get_freqs_indexes(8, 12)].mean(1)
             actual2 = out.mean(
                 axis=(0, -1))[:, self.__get_freqs_indexes(28, 32)].mean(1)
-            np.testing.assert_allclose(actual1, val*np.ones_like(actual),
+            np.testing.assert_allclose(actual1, val * np.ones_like(actual),
                                        atol=atol)
-            np.testing.assert_allclose(actual2, val*np.ones_like(actual),
+            np.testing.assert_allclose(actual2, val * np.ones_like(actual),
                                        atol=atol)
 
     def test_conn_spec(self,):
@@ -99,7 +99,7 @@ class TestConnSpec:
             actual = out.mean(dim=("trials", "times")).sel(
                 freqs=slice(28, 32)).mean("freqs")
             np.testing.assert_allclose(
-                actual, 0.80*np.ones_like(actual), atol=0.1)
+                actual, 0.80 * np.ones_like(actual), atol=0.1)
 
             # 2. Compare with no stationary signal
             x = self.__get_signal(stationary=False)
@@ -116,9 +116,9 @@ class TestConnSpec:
                 val = 0.8
             else:
                 val = 0.9
-            np.testing.assert_allclose(actual_1, val*np.ones_like(actual_1),
+            np.testing.assert_allclose(actual_1, val * np.ones_like(actual_1),
                                        atol=0.1)
-            np.testing.assert_allclose(actual_2, val*np.ones_like(actual_2),
+            np.testing.assert_allclose(actual_2, val * np.ones_like(actual_2),
[…]
