### 2022-01-04 ad03919
Move stat functions from core to stats
