+++ docs/source/api/api_statistics.rst
@@ -9,7 +9,17 @@ Statistics
    :no-members:
    :no-inherited-members:
 
-Random effect (rfx)
+Non-parametric statistics
++++++++++++++++++++++++++
+
+.. autosummary::
+   :toctree: generated/
+
+   permute_mi_vector
+   permute_mi_trials
+   bootstrap_partitions
+
+Random-effect (rfx)
 +++++++++++++++++++
 
 .. autosummary::
