### 2022-01-05 8189622
Function for converting distribution to confidence intervals
