### 2022-01-05 fc58475
Avoid stacking variances, take the max asap
