### 2022-01-05 fbce222
confidence_interval method in the WfMi + example
