+++ frites/workflow/wf_mi.py
@@ -7,11 +7,11 @@ import xarray as xr
 from mne.utils import ProgressBar
 
 from frites.io import set_log_level, logger
-from frites.core import permute_mi_vector
 from frites.workflow.wf_stats import WfStats
 from frites.workflow.wf_base import WfBase
 from frites.estimator import GCMIEstimator, ResamplingEstimator
-from frites.utils import parallel_func, kernel_smoothing
+from frites.utils import parallel_func, kernel_smoothing, nonsorted_unique
+from frites.stats import dist_to_ci, permute_mi_vector, bootstrap_partitions
 
 
 class WfMi(WfBase):
@@ -128,7 +128,7 @@ class WfMi(WfBase):
             # get the randomize version of y
             y_p = permute_mi_vector(
                 y, suj, mi_type=self._mi_type, inference=self._inference,
-                n_perm=n_perm)
+                n_perm=n_perm, random_state=random_state)
             # run permutations using the randomize regressor
             _mi_p = parallel(p_fun(x, y_p[p], **kw_mi) for p in range(n_perm))
             _mi_p = np.asarray(_mi_p)
@@ -367,6 +367,204 @@ class WfMi(WfBase):
             conj[k], conj_ss[k] = v, v
 
         return conj_ss, conj
+
+    def confidence_interval(self, dataset, ci=95, n_boots=200, rfx_es='mi',
+                            n_jobs=-1, random_state=None, verbose=None):
+        """Estimate the empirical confidence interval.
+
+        Parameters
+        ----------
+        dataset : :class:`frites.dataset.DatasetEphy`
+            A dataset instance. If the workflow has already been fitted, then
+            this parameter can remains to None.
+        ci : float, list | 95
+            Confidence level to use in percentage. Use either a single float
+            (e.g. 95, 99 etc.) or a list of floats (e.g. [95, 99])
+        n_boots : int | 200
+            Number of resampling to perform
+        rfx_es : {'mi', 'tvalues'}
+            For the RFX model, specify whether the confidence interval has to
+            be estimated on a measure of effect size in bits ('mi') or on
+            second-level t-test ('tvalues')
+        n_jobs : int | -1
+            Number of jobs to use for parallel computing (use -1 to use all
+            jobs)
+        random_state : int | None
+            Fix the random state of the machine (use it for reproducibility).
+            If None, a random state is randomly assigned.
+
+        Returns
+        -------
+        ci : xr.DataArray
+            Confidence interval array of shape (n_ci, 2, n_times, n_roi) where
+            n_ci describe the number of confidence levels define with the
+            input parameter `ci`, and 2 represents the lower and upper bounds
+            of the confidence interval
+        """
+        set_log_level(verbose)
+
+        # check inputs
+        if isinstance(ci, (int, float)):
+            ci = [ci]
+        assert isinstance(ci, (list, tuple, np.ndarray))
+        assert isinstance(n_boots, int)
+        for k in ci:
+            assert isinstance(k, (int, float)) and (0 < k < 100)
+        assert len(self._mi) and len(self._mi_p), (
+            "You've to lauched the workflow (`fit()`) before being able to "
+            "perform the conjunction analysis.")
+        assert rfx_es in ['mi', 'tvalues']
+
+        """
+        For the RFX, the CI on t-values is estimated at the second level i.e.
+        by bootstraping the subjects. I'm not sure why, but CI at the trial
+        level (i.e. resampling trials per subjects) leads to t-values bellow
+        the real effect size.
+        """
+        if (self._inference == 'rfx') and (rfx_es == 'tvalues'):
+            from frites.config import CONFIG
[…]
