### 2022-01-05 265c3a8
Tests new non-parametric functions
