+++ .github/workflows/pr.yml
@@ -25,6 +25,8 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install -e .['full']
+          # install dev version of mne-connectivity
+          pip install -U https://github.com/mne-tools/mne-connectivity/archive/main.zip
 
       - name: Test with pytest 🔧
         run: |
