### 2022-01-14 c86b19f
conn_net for computing the net connectivity
