### 2022-02-07 867b604
Create matplotlib style for frites
