### 2022-02-14 a0d0182
Add function for generating connectivity links + tests
