+++ frites/conn/__init__.py
@@ -23,5 +23,5 @@ from .conn_spec import conn_spec  # noqa
 # connectivity utility functions
 from .conn_fcd_corr import conn_fcd_corr  # noqa
 from .conn_sliding_windows import define_windows, plot_windows  # noqa
-from .conn_utils import (conn_get_pairs, conn_reshape_undirected,  # noqa
+from .conn_utils import (conn_get_pairs, conn_links, conn_reshape_undirected,  # noqa
                          conn_reshape_directed, conn_ravel_directed, conn_net)
