### 2022-02-14 78bf364
Improve transfer entropy function
