+++ frites/conn/conn_io.py
@@ -7,12 +7,13 @@ import mne
 from frites.io import set_log_level, logger
 from frites.config import CONFIG
 from frites.dataset import SubjectEphy
+from frites.conn.conn_utils import conn_links
 
 
 def conn_io(data, times=None, roi=None, y=None, sfreq=None, agg_ch=False,
-            win_sample=None, pairs=None, freqs=None, foi=None, sm_times=None,
-            sm_freqs=None, directed=False, block_size=None, name=None,
-            sort=True, verbose=None):
+            win_sample=None, freqs=None, foi=None, sm_times=None,
+            sm_freqs=None, block_size=None, name=None,
+            kw_links=dict(), verbose=None):
     """Prepare connectivity variables.
 
     Parameters
@@ -39,8 +40,6 @@ def conn_io(data, times=None, roi=None, y=None, sfreq=None, agg_ch=False,
         finish. You can use the function :func:`frites.conn.define_windows`
         to define either manually either sliding windows. If None, the entire
         time window is used instead.
-    pairs : array_like | None
-        Pairs of contacts
     freqs : array_like | None
         Vector of frequencies
     foi : array_like | None
@@ -51,17 +50,11 @@ def conn_io(data, times=None, roi=None, y=None, sfreq=None, agg_ch=False,
         Number of points to consider for the temporal smoothing in seconds
     sm_freqs : float
         Number of points to consider for the frequency smoothing in Hz
-    directed : bool | False
-        Get either the list of pairs of brain regions for undirected (upper
-        triangle) or directed (off-diagonal elements)
     agg_ch : bool | False
         In case there are multiple electrodes, channels, contacts or sources
         inside a brain region, specify how the data has to be aggregated.
     block_size : int | None
         Number of blocks of trials to process at once.
-    sort : bool | True
-        Specify whether roi names should be sorted (True for undirected
-        measures) or not (False for directed measures)
 
     Returns
     -------
@@ -97,39 +90,20 @@ def conn_io(data, times=None, roi=None, y=None, sfreq=None, agg_ch=False,
 
     # _________________________________ SPACE _________________________________
     # get indices of pairs of (group) regions
-    if (pairs is None):
-        # deal with channel aggregation
-        if agg_ch:  # for multi-variate computations
-            gp = pd.DataFrame({'roi': roi}).groupby('roi', sort=False).groups
-            roi_gp = np.array(list(gp.keys()))
-            roi_idx = [list(k) for k in gp.values()]
-        else:
-            roi_gp, roi_idx = roi, np.arange(len(roi)).reshape(-1, 1)
-
-        # get pairs for directed / undirected conn
-        if directed:
-            x_s, x_t = np.where(~np.eye(len(roi_gp), dtype=bool))
-        else:
-            x_s, x_t = np.triu_indices(len(roi_gp), k=1)
+    if agg_ch:  # for multi-variate computations
+        gp = pd.DataFrame({'roi': roi}).groupby('roi', sort=False).groups
+        roi_gp = np.array(list(gp.keys()))
+        roi_idx = [list(k) for k in gp.values()]
     else:
-        assert isinstance(pairs, np.ndarray)
-        assert (pairs.ndim == 2) and (pairs.shape[1] == 2)
-        x_s, x_t = pairs[:, 0], pairs[:, 1]
         roi_gp, roi_idx = roi, np.arange(len(roi)).reshape(-1, 1)
+
+    # get connectivity links
+    (x_s, x_t), roi_p = conn_links(roi_gp, **kw_links)
 
[…]
