+++ docs/source/api/api_connectivity.rst
@@ -34,8 +34,8 @@ Reshaping connectivity outputs
    conn_reshape_directed
    conn_ravel_directed
    conn_net
-   conn_get_pairs
    conn_links
+   conn_get_pairs
 
 Metrics to apply on FC
 ======================
