### 2022-02-14 ad8a2b0
More control when computing TE
