### 2022-02-15 8d69d77
Add sfreq to attributes
