### 2022-02-17 97064b8
Frites style integration
