+++ examples/statistics/plot_wf_mi_stats_compare_rfx.py
@@ -4,16 +4,17 @@ Compare between-subjects statistics when computing mutual information
 
 This example illustrates how to define and run a workflow for computing
 mutual information and evaluate stastitics. Inference are made within-subjects
-(random effect = rfx). 
+(random effect = rfx).
 """
 import numpy as np
 
 from frites.simulations import sim_multi_suj_ephy, sim_mi_cc
 from frites.dataset import DatasetEphy
 from frites.workflow import WfMi
+from frites import set_mpl_style
 
 import matplotlib.pyplot as plt
-plt.style.use('seaborn-white')
+set_mpl_style()
 
 
 ###############################################################################
@@ -84,6 +85,7 @@ Use either :
 """
 cluster_th = None  # {float, None, 'tfce'}
 
+plt.figure(figsize=(10, 8))
 for mcp in mcps:
     print(f'-> RFX / MCP={mcp}')
     mi, pvalues = wf.fit(dt, mcp=mcp, cluster_th=cluster_th, **kw)
@@ -99,7 +101,7 @@ plt.legend()
 
 plt.subplot(211)
 plt.plot(time, mi.squeeze())
-plt.xlabel('Time (s)'), plt.ylabel("Bits")
+plt.ylabel("Bits")
 plt.title("Mutual information I(C; C)")
 plt.autoscale(tight=True)
 
