+++ frites/conn/conn_dfc.py
@@ -61,7 +61,7 @@ def conn_dfc(data, win_sample=None, times=None, roi=None, agg_ch=False,
         jobs). The parallel loop is set at the pair level.
     kw_links : dict | {}
         Additional arguments for selecting links to compute are passed to the
-        function conn_links
+        function :func:`frites.conn.conn_links`
 
     Returns
     -------
