+++ _downloads/b5a5b61f87dd28805f5fe3b2e5ca1f01/plot_wf_mi_ccd.py
@@ -13,9 +13,10 @@ to partial correlation. For further details, see Ince et al., 2017
 from frites.simulations import sim_multi_suj_ephy, sim_mi_ccd
 from frites.dataset import DatasetEphy
 from frites.workflow import WfMi
+from frites import set_mpl_style
 
 import matplotlib.pyplot as plt
-plt.style.use('seaborn-white')
+set_mpl_style()
 
 
 ###############################################################################
