+++ _downloads/eb0def8ff17c49846bf519a0c946ac95/plot_wf_mi_combine.py
@@ -13,9 +13,10 @@ import xarray as xr
 from frites.simulations import sim_local_cd_ms
 from frites.dataset import DatasetEphy
 from frites.workflow import WfMi, WfMiCombine
+from frites import set_mpl_style
 
 import matplotlib.pyplot as plt
-plt.style.use('seaborn-white')
+set_mpl_style()
 
 
 ###############################################################################
