+++ _downloads/31a26a1996e5a607840326d75d11b163/plot_conn.py
@@ -13,9 +13,10 @@ from itertools import product
 from frites.simulations import sim_multi_suj_ephy
 from frites.dataset import DatasetEphy
 from frites.workflow import WfConnComod
+from frites import set_mpl_style
 
 import matplotlib.pyplot as plt
-plt.style.use('seaborn-white')
+set_mpl_style()
 
 
 ###############################################################################
