### 2022-02-17 5092a35
Fix position of the MANIFEST.in (how stupid i'm)
