+++ MANIFEST.in
@@ -3,5 +3,5 @@ include README.rst
 include LICENSE
 include requirements.txt
 
-# Add url path :
-include frites/data/frites.mplstyle
+# include files in data/
+graft frites/data
