### 2022-03-03 59e15e6
Downsampling function
