+++ docs/source/api/api_utils.rst
@@ -15,6 +15,7 @@ Time-series processing
 .. autosummary::
    :toctree: generated/
 
+   downsample
    acf
 
 Data smoothing
