+++ docs/source/overview/ovw_whatsnew.rst
@@ -14,6 +14,7 @@ New Features
 * New function :func:`frites.conn.conn_net` for computing the net connectivity (:commit:`c86b19f0`)
 * New function :func:`frites.set_mpl_style` for example styles
 * New function :func:`frites.conn.conn_links` for generating connectivity links (:commit:`a0d0182d1`)
+* New function :func:`frites.utils.downsample` for down-sampling DataArray (:commit:`59e15e67`)
 
 v0.4.1
 ------
