+++ .github/workflows/pr.yml
@@ -14,7 +14,7 @@ jobs:
 
     steps:
       - name: Checkout 🛎️
-        uses: actions/checkout@v2.4.0
+        uses: actions/checkout@v3
 
       - name: Set up Python ${{ matrix.python-version }} 🔧
         uses: actions/setup-python@v3
