### 2022-03-25 f6e80e9
Extend CorrEstimator with vector and tensor-based pearson and spearman correlations
