+++ frites/estimator/est_corr.py
@@ -1,5 +1,6 @@
 """Correlation based estimators."""
 import numpy as np
+from scipy.stats import spearmanr
 
 from frites.estimator.est_mi_base import BaseMIEstimator
 from frites.utils import jit
@@ -11,13 +12,42 @@ class CorrEstimator(BaseMIEstimator):
 
     This estimator can be used to estimate the correlation between two
     continuous variables (mi_type='cc').
+
+    Parameters
+    ----------
+    method : {'pearson', 'spearman'}
+        Use either the Pearson correlation or the rank-based Spearman
+        correlation
+    implementation : {'vector', 'tensor'}
+        Specify whether to use the traditional vector-based implementation or
+        the tensor-based implementation (usually faster)
     """
 
-    def __init__(self, verbose=None):
+    def __init__(self, method='pearson', implementation='vector',
+                 verbose=None):
         """Init."""
         self.name = 'Correlation-based Estimator'
-        super(CorrEstimator, self).__init__(mi_type='cc', verbose=verbose)
-        self._core_fun = correlate
+
+        # implementation selection
+        _methods = {
+            'pearson': {
+                'tensor': ten_pearson,
+                'vector': vec_pearson,
+                # 'numba': nb_pearson
+            },
+            'spearman': {
+                'tensor': ten_spearman,
+                'vector': vec_spearman,
+                # 'numba': vec_spearman
+            }
+        }
+        self._method = method
+        self._implementation = implementation
+        self._core_fun = _methods[method][implementation]
+        # additional string for the description
+        add_str = f", method={method}, implementation={implementation}"
+        super(CorrEstimator, self).__init__(mi_type='cc', verbose=verbose,
+                                            add_str=add_str)
         # update internal settings
         settings = dict(mi_type='cc', core_fun=self._core_fun.__name__)
         self.settings.merge([settings])
@@ -76,22 +106,49 @@ class CorrEstimator(BaseMIEstimator):
                 n_var, n_mv, _ = x.shape
                 y = np.tile(y, (n_var, 1, 1))
 
-            # types checking
-            if x.dtype != np.float32:
-                x = x.astype(np.float32, copy=False)
-            if y.dtype != np.float32:
-                y = y.astype(np.float32, copy=False)
-            if categories.dtype != np.int32:
-                categories = categories.astype(np.int32, copy=False)
+            # numba related changes
+            if (self._implementation == 'numba') and (
+                self._method == 'pearson'):
+                if x.dtype != np.float32:
+                    x = x.astype(np.float32, copy=False)
+                if y.dtype != np.float32:
+                    y = y.astype(np.float32, copy=False)
+                if categories.dtype != np.int32:
+                    categories = categories.astype(np.int32, copy=False)
+
+            # change flags (compatibility with multi-core processing)
+                x.flags.writeable = True
+                y.flags.writeable = True
+                categories.flags.writeable = True
+                x = np.ascontiguousarray(x)
+                y = np.ascontiguousarray(y)
+                categories = np.ascontiguousarray(categories)
 
             return core_fun(x, y, categories)
 
         return estimator
 
 
-@jit("f4[:, :](f4[:,:,:], f4[:,:,:], i4[:])")
-def correlate(x, y, categories):
-    """3D correlation."""
+###############################################################################
[…]
