### 2022-03-25 e02e1a7
Patch for recent version of MNE
