+++ frites/conn/conn_tf.py
@@ -72,15 +72,30 @@ def _tf_decomp(data, sf, freqs, mode='morlet', n_cycles=7.0, mt_bandwidth=None,
             assert len(freqs) == len(n_cycles) == len(mt_bandwidth)
             out = []
             for f_c, n_c, mt in zip(freqs, n_cycles, mt_bandwidth):
-                out += [tfr_array_multitaper(
+                _out = tfr_array_multitaper(
                     data, sf, [f_c], n_cycles=float(n_c), time_bandwidth=mt,
-                    output='complex', decim=decim, n_jobs=n_jobs, **kw_mt)]
+                    output='complex', decim=decim, n_jobs=n_jobs, **kw_mt
+                )
+
+                # recent version of mne allows to return the TF decomposition
+                # with the additional n_tapers dimension. This patch takes the
+                # mean over the tapers
+                if _out.ndim == 5: _out = _out.mean(2)
+                out.append(_out)
+
+            # stack everything
             out = np.stack(out, axis=2).squeeze()
         elif isinstance(mt_bandwidth, (type(None), int, float)):
             out = tfr_array_multitaper(
                 data, sf, freqs, n_cycles=n_cycles,
                 time_bandwidth=mt_bandwidth, output='complex', decim=decim,
                 n_jobs=n_jobs, **kw_mt)
+
+            # recent version of mne allows to return the TF decomposition
+            # with the additional n_tapers dimension. This patch takes the
+            # mean over the tapers
+            if out.ndim == 5: out = out.mean(2)
+
     else:
         raise ValueError('Method should be either "morlet" or "multitaper"')
 
