### 2022-03-28 abea667
CONFIG dict do not allow anymore to add new keys but still allow to update values
