### 2022-03-28 aab68e1
rfx_center and rfx_sigma input parameters to better control RFX
