### 2022-03-31 87c762c
intra and inter hemispheric connection selection in conn_links
