### 2022-04-06 74d2cc1
Allow to select both / intra / inter roi relationships
