+++ .github/workflows/test_doc.yml
@@ -14,7 +14,7 @@ jobs:
 
     steps:
     - name: Checkout 🛎️
-      uses: actions/checkout@v2.4.0
+      uses: actions/checkout@v3.0.0
 
     - name: Set up Python ${{ matrix.python-version }} 🔧
       uses: actions/setup-python@v3
@@ -35,7 +35,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - name: Checkout 🛎️
-        uses: actions/checkout@v2.4.0
+        uses: actions/checkout@v3.0.0
 
       - name: Set up Python 🔧
         uses: actions/setup-python@v3
