+++ docs/source/overview/ovw_whatsnew.rst
@@ -3,6 +3,9 @@
 What's new
 ==========
 
+v0.4.3
+------
+
 v0.4.2
 ------
 
