### 2022-05-11 8d7e1fb
Allows to use brain region names as sources / targets
