+++ docs/source/overview/ovw_whatsnew.rst
@@ -6,6 +6,10 @@ What's new
 v0.4.3
 ------
 
+New Features
+++++++++++++
+* :func:`frites.conn.conn_links` now accepts to use brain region names as source or target seeds (:commit:`8d7e1fb6`)
+
 v0.4.2
 ------
 
