### 2022-05-22 f2ff5f6
Fix computations of covgc in parallel
