### 2022-05-22 940362e
Include Deco's normalization to the covgc
