+++ frites/conn/conn_covgc.py
@@ -5,7 +5,7 @@ import xarray as xr
 from frites.conn import conn_io
 from frites.io import set_log_level, logger, check_attrs
 from frites.config import CONFIG
-from frites.core.gcmi_nd import cmi_nd_ggg
+from frites.core.gcmi_nd import cmi_nd_ggg, mi_nd_gg
 from frites.core.copnorm import copnorm_nd
 from frites.utils import parallel_func
 
@@ -18,7 +18,6 @@ from frites.utils import parallel_func
 
 
 LOG2 = np.log(2)
-kw_gc = {k: v for k, v in CONFIG['KW_GCMI'].items()}  # patch for parallel computing
 
 
 def entr(xy):
@@ -43,7 +42,7 @@ def entr(xy):
     return h
 
 
-def _covgc(d_s, d_t, ind_tx, t0, norm=False):
+def _covgc(d_s, d_t, ind_tx, t0, norm, **kw_gc):
     """Compute the covGC for a single pair.
 
     This function computes the covGC for a single pair, across multiple trials,
@@ -100,23 +99,23 @@ def _covgc(d_s, d_t, ind_tx, t0, norm=False):
             # -----------------------------------------------------------------
             # Normalisation of GC (Deco et al, Nature Human Beh, 2021)
             # -----------------------------------------------------------------
-            # if norm:  # this part requires more work
-            #     # H(Y_i+1)
-            #     y0 = np.expand_dims(y[0], axis=0)
-            #     hy = entr(y0)
-            #     # H(X_i+1)
-            #     x0 = np.expand_dims(x[0], axis=0)
-            #     hx = entr(x0)
-            #     # I(Y_i+1; Y_i) internal predictability of Y
-            #     mi_yi1yi = hy - hycy
-            #     # I(X_i+1; X_i) internal predictability of X
-            #     mi_xi1xi = hx - hxcx
-            #     # gc(pairs(:,1) -> pairs(:,2))
-            #     gc[n_tr, n_ti, 0] = gc[n_tr, n_ti, 0] / (
-            #         mi_yi1yi + gc[n_tr, n_ti, 0])
-            #     # gc(pairs(:,2) -> pairs(:,1))
-            #     gc[n_tr, n_ti, 1] = gc[n_tr, n_ti, 1] / (
-            #         mi_xi1xi + gc[n_tr, n_ti, 1])
+            if norm:
+                # H(Y_i+1)
+                y0 = np.expand_dims(y[0], axis=0)
+                hy = entr(y0)
+                # H(X_i+1)
+                x0 = np.expand_dims(x[0], axis=0)
+                hx = entr(x0)
+                # I(Y_i+1; Y_i) internal predictability of Y
+                mi_yi1yi = hy - hycy
+                # I(X_i+1; X_i) internal predictability of X
+                mi_xi1xi = hx - hxcx
+                # gc(pairs(:,1) -> pairs(:,2))
+                gc[n_tr, n_ti, 0] = gc[n_tr, n_ti, 0] / (
+                    mi_yi1yi + gc[n_tr, n_ti, 0])
+                # gc(pairs(:,2) -> pairs(:,1))
+                gc[n_tr, n_ti, 1] = gc[n_tr, n_ti, 1] / (
+                    mi_xi1xi + gc[n_tr, n_ti, 1])
 
     return gc
 
@@ -128,7 +127,7 @@ def _covgc(d_s, d_t, ind_tx, t0, norm=False):
 ###############################################################################
 
 
-def _gccovgc(d_s, d_t, ind_tx, t0):
+def _gccovgc(d_s, d_t, ind_tx, t0, norm, **kw_gc):
     """Compute the Gaussian-Copula based covGC for a single pair.
 
[…]
