+++ docs/source/refs.bib
@@ -244,3 +244,14 @@
   pages = {126--136},
   file = {Combrisson et Jerbi - 2015 - Exceeding chance level by chance The caveat of th.pdf:/home/etienne/Zotero/storage/ARX2855B/Combrisson et Jerbi - 2015 - Exceeding chance level by chance The caveat of th.pdf:application/pdf}
 }
+
+@article{deco2021revisiting,
+  title={Revisiting the global workspace orchestrating the hierarchical organization of the human brain},
+  author={Deco, Gustavo and Vidaurre, Diego and Kringelbach, Morten L},
+  journal={Nature human behaviour},
+  volume={5},
+  number={4},
+  pages={497--511},
+  year={2021},
+  publisher={Nature Publishing Group}
+}
