### 2022-05-23 3a51253
New function for computing confidence interval, sd and sem on mean mi
