+++ frites/stats/stats_nonparam.py
@@ -1,8 +1,11 @@
 """Utility functions for stat evaluation."""
 import numpy as np
+import xarray as xr
 
 from frites.utils import nonsorted_unique
 from frites.dataset.ds_utils import multi_to_uni_conditions
+from frites.io.io_attributes import check_attrs
+from frites.io import logger, set_log_level
 
 
 def permute_mi_vector(y, suj, mi_type='cc', inference='rfx', n_perm=1000,
@@ -201,3 +204,121 @@ def dist_to_ci(dist, cis=[99], inference='ffx', rfx_es='mi', pop_mean=None):
         x_ci[n_ci, 1, :] = np.percentile(dist, (100. - half_alpha), axis=0)
 
     return x_ci
+
+
+def confidence_interval(data, axis=0, cis=95, n_boots=200, random_state=None,
+                        fcn=np.mean, verbose=None):
+    """Compute the confidence interval of repeated measurements.
+
+    Parameters
+    ----------
+    data : array_like
+        Numpy array (or xarray.DataArray) of data
+    axis : int | 0
+        Axis along which to compute the confidence interval
+    cis : int, list | 95
+        Integer or list of confidence levels to extract. This input also
+        supports computing standard deviation ('sd') and / or standard error on
+        the mean ('sem')
+    n_boots : int | 200
+        Number of bootstraps to perform
+    random_state : int | None
+        Fix the random state of the machine (use it for reproducibility). If
+        None, a random state is randomly assigned.
+    fcn : function | np.mean
+        Summary statistics function. By default, the mean is used.
+
+    Returns
+    -------
+    ci : array_like
+        Array of confidence intervals of shape (n_ci, 2, ...) where n_ci
+        refers to the number of desired confidence intervals (see input `cis`)
+        and 2 refers to the lower and upper bounds.
+    """
+    set_log_level(verbose)
+    # ---------------------------------- I/O ----------------------------------
+    if isinstance(cis, (int, float, str)):
+        cis = [cis]
+    assert isinstance(cis, (list, tuple, np.ndarray))
+    assert isinstance(n_boots, int)
+    need_ci = np.any([isinstance(k, (int, float)) for k in cis])
+    logger.info(f"    Estimating CI (cis={cis}, axis={axis}, "
+                f"n_boots={n_boots}, random_state={random_state})")
+
+    # ------------------------------- DATAARRAY -------------------------------
+    if isinstance(data, xr.DataArray):
+        if isinstance(axis, str):
+            axis = data.get_axis_num(axis)
+        dims = [d for n_d, d in enumerate(data.dims) if n_d != axis]
+        coords = [data[d].data for d in dims]
+        attrs = data.attrs
+        attrs.update(n_boots=n_boots, random_state=random_state,
+                     fcn=fcn.__name__)
+        attrs = check_attrs(attrs)
+        name = 'CI' if data.name is None else data.name + '_CI'
+        x = data.data
+    else:
+        x = data
+
+    # ------------------------------- BOOSTRAPS -------------------------------
+    if need_ci:
+        # compute summary statistics
+        part = bootstrap_partitions(x.shape[axis], n_partitions=n_boots,
+                                    random_state=random_state)
+        x_ci = []
+        for k in range(n_boots):
+            sl = [slice(None)] * x.ndim
+            sl[axis] = part[k]
+            x_ci.append(fcn(x[tuple(sl)], axis=axis))
+        x_ci = np.stack(x_ci)
+
[…]
