### 2022-05-24 169df5a
Add scikit-learn to the fixed dependencies
