+++ requirements.txt
@@ -5,3 +5,4 @@ joblib
 xarray
 netCDF4
 h5netcdf
+scikit-learn
