+++ docs/source/overview/ovw_acknowledgments.rst
@@ -1,0 +1,1 @@
+This open source software code was developed in part in the Human Brain Project, funded from the European Union’s Horizon 2020 Framework Programme for Research and Innovation under Specific Grant Agreements No. 945539 (Human Brain Project SGA3) and in the ANR PRC project “CausaL” (ANR-18-CE28-0016).
