### 2022-05-25 75c2311
Fix computations of SEM (wrong axis)
