### 2022-06-30 a1ecfc3
Fix url links in ovw_frites
