### 2022-06-30 df0af9f
Add performance section to the doc and an example
