+++ docs/source/overview/index.rst
@@ -9,6 +9,7 @@ This section introduces the basis for starting analyzing your data with Frites.
     ovw_goals
     ovw_methods
     ovw_frites
+    ovw_perf
     ovw_help
     ovw_cite
     ovw_community
