+++ paper/paper.md
@@ -65,7 +65,7 @@ package for cognitive (task-related) brain network inference, which combines in 
 
 The package integrates a non-parametric permutation-based statistical framework to perform group-level inferences on non-negative measures of information. The toolbox includes different methods that cope with multiple-comparison correction problems, such as test- and cluster-wise p-value corrections. The implemented framework supports both fixed- and random-effect models to adapt to inter-individuals and inter-sessions variability [@Combrisson:2021]. 
 
-`Frites` is optimally designed for the analysis of continuous and multi-channel neurophysiological data, encompassing recordings with either uniform spatial sampling (e.g., M/EEG data) and spatially sparse intracranial recordings, such as intracranial EEG or Local Field Potentials (LFPs). The package supports standard [`NumPy`](https://numpy.org/) array inputs [@Harris:2020], objects from the [`MNE-Python`](https://mne.tools/stable/index.html)  software [@Gramfort:2013], but also multi-dimensional labelled  [`Xarray`](http://xarray.pydata.org/en/stable/) objects [@Hoyer:2017].
+`Frites` is suitable for the analysis of continuous and multi-channel neurophysiological data, encompassing recordings with either uniform spatial sampling (e.g., M/EEG data) and spatially sparse intracranial recordings, such as intracranial EEG or Local Field Potentials (LFPs). The package supports standard [`NumPy`](https://numpy.org/) array inputs [@Harris:2020], objects from the [`MNE-Python`](https://mne.tools/stable/index.html)  software [@Gramfort:2013], but also multi-dimensional labelled  [`Xarray`](http://xarray.pydata.org/en/stable/) objects [@Hoyer:2017].
 
 In order to facilitate automated and efficient usage, `Frites` provides a set of high-level workflows that integrate several analysis steps from information-based estimation to network-level statistical inference.
 
