+++ paper/paper.bib
@@ -114,12 +114,22 @@
     pages = {e1008302}
 }
 
+
 @article{Combrisson:2021,
-  title={Group-level inference of information-based measures for the analyses of cognitive brain networks from neurophysiological data},
-  author={Combrisson, Etienne and Allegra, Michele and Basanisi, Ruggero and Ince, Robin AA and Giordano, Bruno and Bastin, Julien and Brovelli, Andrea},
-  journal={bioRxiv},
-  year={2021},
-  publisher={Cold Spring Harbor Laboratory}
+    title = {Group-level inference of information-based measures for the analyses of cognitive brain networks from neurophysiological data},
+    volume = {258},
+    issn = {1053-8119},
+    url = {https://www.sciencedirect.com/science/article/pii/S1053811922004669},
+    doi = {10.1016/j.neuroimage.2022.119347},
+    language = {en},
+    urldate = {2022-06-30},
+    journal = {NeuroImage},
+    author = {Combrisson, Etienne and Allegra, Michele and Basanisi, Ruggero and Ince, Robin A. A. and Giordano, Bruno L. and Bastin, Julien and Brovelli, Andrea},
+    month = sep,
+    year = {2022},
+    keywords = {Cluster-based, Group-level statistics, Information-based measures, Neurophysiology, Non-parametric, python, Reproducibility},
+    pages = {119347},
+    file = {ScienceDirect Snapshot:/home/etienne/Zotero/storage/CVAP7X27/S1053811922004669.html:text/html},
 }
 
 @article{Wollstadt:2019,
