+++ docs/source/refs.bib
@@ -255,3 +255,21 @@
   year={2021},
   publisher={Nature Publishing Group}
 }
+
+
+@article{combrisson_group-level_2022,
+  title = {Group-level inference of information-based measures for the analyses of cognitive brain networks from neurophysiological data},
+  volume = {258},
+  issn = {1053-8119},
+  url = {https://www.sciencedirect.com/science/article/pii/S1053811922004669},
+  doi = {10.1016/j.neuroimage.2022.119347},
+  language = {en},
+  urldate = {2022-06-30},
+  journal = {NeuroImage},
+  author = {Combrisson, Etienne and Allegra, Michele and Basanisi, Ruggero and Ince, Robin A. A. and Giordano, Bruno L. and Bastin, Julien and Brovelli, Andrea},
+  month = sep,
+  year = {2022},
+  keywords = {Cluster-based, Group-level statistics, Information-based measures, Neurophysiology, Non-parametric, python, Reproducibility},
+  pages = {119347},
+  file = {Combrisson et al. - 2022 - Group-level inference of information-based measure.pdf:/home/etienne/Zotero/storage/ZRQPHPVA/Combrisson et al. - 2022 - Group-level inference of information-based measure.pdf:application/pdf;ScienceDirect Snapshot:/home/etienne/Zotero/storage/CVAP7X27/S1053811922004669.html:text/html},
+}
