+++ paper/paper.bib
@@ -115,7 +115,7 @@
 }
 
 
-@article{Combrisson:2021,
+@article{Combrisson:2022,
     title = {Group-level inference of information-based measures for the analyses of cognitive brain networks from neurophysiological data},
     volume = {258},
     issn = {1053-8119},
