+++ paper/paper.md
@@ -44,38 +44,26 @@ bibliography: paper.bib
 ---
 
 # Summary
-
-The field of cognitive computational neuroscience addresses open questions regarding
-the complex relation between cognitive functions and the dynamic coordination of neural
-activity over large-scale and hierarchical brain networks. State-of-the-art approaches
-involve the characterization of brain regions and inter-areal interactions that participate
-in cognitive processes [@Battaglia:2020]. More precisely, the study of cognitive
-brain networks underlies linking brain data to experimental variables, such as sensory
-stimuli or behavioral responses.
-
-Information-based measures, such as information theoretic quantities, machine-learning models or measures of distances, currently provide ideal tools for quantifying the coupling between brain signals, and the link between brain network’s activity and task variables. Nevertheless, progress is limited by the lack of neuroinformatics tools that combine methods for the estimate of information-based measures from neural data and the assessment of their statistical relevance at the population level. `Frites` provides such an integrated framework and it is optimally developed for the discovery of cognitive brain networks from multi-channel neurophysiological datasets.
+The field of cognitive computational neuroscience addresses open questions regarding the complex relation between cognitive functions and the dynamic coordination of neural activity over large-scale and hierarchical brain networks. State-of-the-art approaches involve the characterization of brain regions and inter-areal interactions that participate in cognitive processes [@Battaglia:2020]. More precisely, the study of cognitive brain networks underlies linking brain data to experimental variables, such as sensory stimuli or behavioral responses. The relation between the neural activity and external variables might take complex forms (e.g. non-linear relationships) with strong variations across brain regions and participants. Therefore, powerful measures of information are required to detect complex relation and the statistical relevance at the population level should be able to adapt to the inter subject variability. 
 
 # Statement of need
+[`Frites`](https://brainets.github.io/frites) (_Framework for Information Theoretical analysis of Electrophysiological data and Statistics_) is an open-source Python package for the discovery of cognitive brain networks. It combines in a single framework information-based analyses of neurophysiological recordings and group-level statistical inference that can adapt to the inherent variability across participants. The functionalities are wrapped into high-level workflows accessible to users with little programming knowledge.
 
-[`Frites`](https://brainets.github.io/frites) (_Framework for Information
-Theoretical analysis of Electrophysiological data and Statistics_) is a pure Python
-package for cognitive (task-related) brain network inference, which combines in a single framework information-based analyses of neurophysiological data and group-level statistical inference.
-
-`Frites` is equipped with a set of information theoretic tools for the analysis of interactions between brain signals and their relation with experimental task-related variables. More precisely, this package can be used to study the relation between local brain activity, cross-frequency coupling [@Combrisson:2020] and inter-areal functional connectivity (FC) with experimental variables (i.e., cognitive tasks). For what concerns FC, the toolbox allows the estimate of dynamic (i.e., time-resolve), undirected (e.g., mutual information) and directed (e.g., Granger causality) functional connectivity (FC) on a single-trial basis [@Brovelli:2015]. The core functions for information measures exploit a `NumPy` tensor-based implementation of the Gaussian Copula Mutual-Information [@Ince:2017]. Nevertheless, the definition of custom estimators, such as kernel methods [@Wollstadt:2019] and [`scikit-learn`](https://scikit-learn.org/stable/) cross-validated classifiers [@Pedregosa:2011], is also supported.
-
[…]
