### 2022-07-06 32d605a
Add parameters to directly mean across trials
