### 2022-07-10 cb58afc
conn_spec supports frequency dependent smoothing
