+++ frites/conn/conn_tf.py
@@ -131,48 +131,20 @@ def _create_kernel(sm_times, sm_freqs, kernel='hanning'):
     kernel : array_like
         Smoothing kernel of shape (sm_freqs, sm_times)
     """
-    scale = isinstance(sm_times, np.ndarray)
+    # frequency dependent kernels
+    if isinstance(sm_times, (np.ndarray, list, tuple)):
+        sm_freqs = 1  # force 1hz smoothing
+        kernels = [_create_kernel(
+            sm, sm_freqs, kernel=kernel) for sm in sm_times]
+        return kernels
 
-    if scale:
-        # I know this piece of code is terrible ='D
-        logger.info("For frequency dependent kernel sm_freqs is not used"
-                    "")
-        # Number of kernels
-        n_kernel = len(sm_times)
-        # Get the size of the biggest kernel
-        max_size = sm_times.max()
-        # Container for the padded kernel
-        s_pad = np.zeros((n_kernel, max_size), dtype=np.float32)
-        # Store kernel for each frequency
-        s = []
-
-        def __pad_kernel(s):
-            for i in range(n_kernel):
-                #  print(f"{s[i]}")
-                pad_size = int(max_size - len(s[i]))
-                # The len(s[i])%2 corrects in case the len is odd
-                s_pad[i, :] = np.pad(
-                    s[i], (pad_size // 2, pad_size // 2 + pad_size % 2))
-            return s_pad
-
+    # frequency independent kernels
     if kernel == 'square':
-        if not scale:
-            return np.full((sm_freqs, sm_times), 1. / (sm_times * sm_freqs))
-        else:
-            for i in range(n_kernel):
-                s += [np.ones(sm_times[i]) / sm_times[i]]
-            # Pad with zeros
-            return __pad_kernel(s)
+        return np.full((sm_freqs, sm_times), 1. / (sm_times * sm_freqs))
     elif kernel == 'hanning':
-        if not scale:
-            hann_t, hann_f = np.hanning(sm_times), np.hanning(sm_freqs)
-            hann = hann_f.reshape(-1, 1) * hann_t.reshape(1, -1)
-            return hann / np.sum(hann)
-        else:
-            for i in range(n_kernel):
-                hann = np.hanning(sm_times[i])
-                s += [hann / np.sum(hann)]
-            return __pad_kernel(s)
+        hann_t, hann_f = np.hanning(sm_times), np.hanning(sm_freqs)
+        hann = hann_f.reshape(-1, 1) * hann_t.reshape(1, -1)
+        return hann / np.sum(hann)
     else:
         raise ValueError(f"No kernel {kernel}")
 
@@ -187,8 +159,8 @@ def _smooth_spectra(spectra, kernel, scale=False, decim=1):
     ----------
     spectra : array_like
         Spectra of shape (..., n_freqs, n_times)
-    kernel : array_like
-        Smoothing kernel of shape (sm_freqs, sm_times)
+    kernel : array_like, list
+        Smoothing kernel  (or list of kernels) of shape (sm_freqs, sm_times)
     decim : int | 1
         Decimation factor to apply after the kernel smoothing
 
@@ -197,18 +169,29 @@ def _smooth_spectra(spectra, kernel, scale=False, decim=1):
     sm_spectra : array_like
         Smoothed spectra of shape (..., n_freqs, n_times)
     """
+    # define axes to use for smoothing
+    axes = -1 if scale else (-2, -1)
+
+    # frequency (in)dependent smoothing
+    if isinstance(kernel, list):
+        for n_k, kern in enumerate(kernel):
+            spectra[..., n_k, :] = __smooth_spectra(
[…]
