+++ paper/paper.bib
@@ -144,3 +144,39 @@
   title = {IDTxl: The Information Dynamics Toolkit xl: a Python package for the efficient analysis of multivariate information dynamics in networks},
   journal = {Journal of Open Source Software}
 }
+
+
+@article{Brovelli:2017,
+    title = {Dynamic {Reconfiguration} of {Visuomotor}-{Related} {Functional} {Connectivity} {Networks}},
+    volume = {37},
+    issn = {0270-6474, 1529-2401},
+    url = {http://www.jneurosci.org/lookup/doi/10.1523/JNEUROSCI.1672-16.2016},
+    doi = {10.1523/JNEUROSCI.1672-16.2016},
+    language = {en},
+    number = {4},
+    urldate = {2019-11-18},
+    journal = {The Journal of Neuroscience},
+    author = {Brovelli, Andrea and Badier, Jean-Michel and Bonini, Francesca and Bartolomei, Fabrice and Coulon, Olivier and Auzias, Guillaume},
+    month = jan,
+    year = {2017},
+    pages = {839--853}
+}
+
+
+@article{Maris:2007,
+    title = {Nonparametric statistical testing of {EEG}- and {MEG}-data},
+    volume = {164},
+    issn = {01650270},
+    url = {https://linkinghub.elsevier.com/retrieve/pii/S0165027007001707},
+    doi = {10.1016/j.jneumeth.2007.03.024},
+    language = {en},
+    number = {1},
+    urldate = {2019-01-14},
+    journal = {Journal of Neuroscience Methods},
+    author = {Maris, Eric and Oostenveld, Robert},
+    month = aug,
+    year = {2007},
+    keywords = {stat},
+    pages = {177--190},
+}
+
