+++ paper/paper.md
@@ -53,9 +53,9 @@ The field of cognitive computational neuroscience addresses open questions regar
 `Frites` is suitable for the analysis of continuous and multi-channel neurophysiological data, encompassing recordings with either uniform spatial sampling (e.g., M/EEG data) and spatially sparse intracranial recordings, such as intracranial EEG or Local Field Potentials (LFPs). The package supports standard [`NumPy`](https://numpy.org/) array inputs [@Harris:2020], objects from the [`MNE-Python`](https://mne.tools/stable/index.html)  software [@Gramfort:2013], but also multi-dimensional labelled  [`Xarray`](http://xarray.pydata.org/en/stable/) objects [@Hoyer:2017].
 
 # Measures of information and group-level statistics
-`Frites` is equipped with a set of information theoretic tools for the analysis of interactions between brain signals and their relation with experimental task-related variables. By default, `Frites` is using the Gaussian Copula Mutual-Information [@Ince:2017] to study the relation between either local brain activity, cross-frequency coupling [@Combrisson:2020] or inter-areal functional connectivity (FC) with experimental variables (i.e., cognitive tasks). For what concerns FC, the toolbox allows the estimate of dynamic (i.e., time-resolve), undirected (e.g., mutual information) and directed (e.g., Granger causality) functional connectivity (FC) on a single-trial basis [@Brovelli:2015]. Nevertheless, the definition of custom estimators, such as kernel methods [@Wollstadt:2019] and [`scikit-learn`](https://scikit-learn.org/stable/) cross-validated classifiers [@Pedregosa:2011], is also supported.
+`Frites` is equipped with a set of information theoretic tools for the analysis of interactions between brain signals and their relation with experimental task-related variables. By default, `Frites` is using the Gaussian Copula Mutual-Information [@Ince:2017] to study the relation between either local brain activity, cross-frequency coupling [@Combrisson:2020] or inter-areal functional connectivity (FC) with experimental variables (i.e., cognitive tasks). For what concerns FC, the toolbox allows the estimate of dynamic (i.e., time-resolve) [@Brovelli:2017], undirected (e.g., mutual information) and directed (e.g., Granger causality) functional connectivity (FC) on a single-trial basis [@Brovelli:2015]. Nevertheless, the definition of custom estimators, such as kernel methods [@Wollstadt:2019] and [`scikit-learn`](https://scikit-learn.org/stable/) cross-validated classifiers [@Pedregosa:2011], is also supported.
 
-For statistical inferences, the package integrates a non-parametric permutation-based statistical framework to perform group-level inferences on non-negative measures of information. The toolbox includes different methods that cope with multiple-comparison correction problems, such as test- and cluster-wise p-value corrections. The implemented framework also supports both fixed- and random-effect models to adapt to inter-individuals and inter-sessions variability [@Combrisson:2022].
+For statistical inferences, the package integrates a non-parametric permutation-based statistical framework [@Maris:2007] to perform group-level inferences on non-negative measures of information. The toolbox includes different methods that cope with multiple-comparison correction problems, such as test- and cluster-wise p-value corrections. The implemented framework also supports both fixed- and random-effect models to adapt to inter-individuals and inter-sessions variability [@Combrisson:2022].
 
 # Workflows
 `Frites` provides a set of workflows that integrate several analysis steps. Those workflows take as inputs the neural data coming from single or multi-participants (or single / multi sessions), estimate the amount of information shared between the brain data and the external variable, at each brain region and time bins, and finally perform network-level statistical inference, corrected for multiple comparisons. Those high-level workflows are particularly useful for users with no or little programming knowledge as all of those steps can be performed in a very few lines of code, with standard inputs like `MNE-Python` epochs.
