+++ paper/paper.bib
@@ -91,7 +91,7 @@
     volume = {35},
     number = {37},
     journal = {The Journal of Neuroscience},
-    author = {Brovelli, X. Andrea and Chicharro, Daniel and Badier, Jean-Michel and Wang, Huifang and Jirsa, Viktor},
+    author = {Brovelli, Andrea and Chicharro, Daniel and Badier, Jean-Michel and Wang, Huifang and Jirsa, Viktor},
     year = {2015},
     pages = {12643--12658}
 }
