### 2022-07-12 c8cccb9
Fix code quality
