+++ paper/paper.md
@@ -13,7 +13,7 @@ tags:
   - fixed-effect ffx
   - random-effect rfx
   - cluster-based statistics
-  - MEG EEG sEEG LFP
+  - MEG EEG sEEG LFPs
   - Granger causality
 authors:
   - name: Etienne Combrisson
@@ -44,16 +44,16 @@ bibliography: paper.bib
 ---
 
 # Summary
-The field of cognitive computational neuroscience addresses open questions regarding the complex relation between cognitive functions and the dynamic coordination of neural activity over large-scale and hierarchical brain networks. State-of-the-art approaches involve the characterization of brain regions and inter-areal interactions that participate in cognitive processes [@Battaglia:2020]. More precisely, the study of cognitive brain networks underlies linking brain data to experimental variables, such as sensory stimuli or behavioral responses. The relation between the neural activity and external variables might take complex forms (e.g. non-linear relationships) with strong variations across brain regions and participants. Therefore, powerful measures of information are required to detect complex relation and the statistical relevance at the population level should be able to adapt to the inter subject variability. 
+The field of cognitive computational neuroscience addresses open questions regarding the complex relation between cognitive functions and the dynamic coordination of neural activity over large-scale and hierarchical brain networks. State-of-the-art approaches involve the characterization of brain regions and inter-areal interactions that participate in cognitive processes [@Battaglia:2020]. More precisely, the study of cognitive brain networks underlies linking local neural activity or interactions between brain regions to experimental variables, such as sensory stimuli or behavioral responses. The relation between the brain data and external variables might take complex forms (e.g. non-linear relationships) with strong variations across brain regions and participants. Therefore, powerful measures of information are required to detect complex relation and the statistical relevance at the population level should be able to adapt to the inter subject variability. 
 
 # Statement of need
-[`Frites`](https://brainets.github.io/frites) (_Framework for Information Theoretical analysis of Electrophysiological data and Statistics_) is an open-source Python package for the discovery of cognitive brain networks i.e. brain regions or interactions between brain regions that are modulated during a cognitive task. It combines in a single framework information-based analyses of neurophysiological recordings and group-level statistical inference that can adapt to the inherent variability across participants. The functionalities are wrapped into high-level workflows accessible to users with little programming knowledge.
+[`Frites`](https://brainets.github.io/frites) (_Framework for Information Theoretical analysis of Electrophysiological data and Statistics_) is an open-source Python package, designed for researchers and students with little programming knowledge and working on the discovery of cognitive brain networks using neurophysiological data. It combines in a single framework information-based analyses of neurophysiological recordings and group-level statistical inference that can adapt to the inherent variability across participants. The measures of information and the group-level statistics functionalities are wrapped into high-level workflows. In addition, `Frites` also includes several measures of functional connectivity (FC) for measuring pairwise interactions between brain regions.
 
 ## Supported neurophysiological recordings and link with other softwares
-`Frites` is suitable for the analysis of continuous and multi-channel neurophysiological data, encompassing recordings with either uniform spatial sampling (e.g., M/EEG data) and spatially sparse intracranial recordings, such as intracranial EEG or Local Field Potentials (LFPs). The package supports standard [`NumPy`](https://numpy.org/) array inputs [@Harris:2020], objects from the [`MNE-Python`](https://mne.tools/stable/index.html) software [@Gramfort:2013], but also multi-dimensional labelled  [`Xarray`](http://xarray.pydata.org/en/stable/) objects [@Hoyer:2017].
[…]
