### 2022-07-12 #3842 @EtienneCmb
Dear @meg-simula and @travisbthomp, please find below our detailed answers to the comments about the paper and about functionality.

# Suggestions for paper

## Summary

> In short, the current summary seems obfuscated and potentially difficult to read for a "non-specialist audience".

We agree with the reviewer that the first version of the summary was not accessible to a non-specialist audience. We simplified the section to be closer to the online documentation. In addition, the `Summary` section now only includes the context and the problematic that Frites tries to solve.

## Statement of need

> The paper does indeed have a section titled 'Statement of Need'. However, this section does not establish a clear need for the software nor does it "clearly state what problems the software is designed to solve and who the target audience is". At least, it does not do these things in a way that can be appreciated by an non-expert in the area, such as myself, who is not currently aware of the state of the art approaches, what problems they could solve better, and why Frites fills any of these gaps.

We improved the *Statement of need* section with the following points :
1. We now clearly state that Frites has been *designed for researchers and students with little programming knowledge and working on the discovery of cognitive brain networks using neurophysiological data*
2. We explain that the software comes with measures of information to find complex relations between brain data and external variables and also that those measures of information are combined with statistics that adapt to the variability across participants

All of those points are then further detailed in the individual sections.

## State of the field

> I cannot find where this has been done in a manner that is decipherable to a non-expert. I would suggest an approach which clearly states the central problem Frites aims to solve and endeavors a comparative discussion to other available tools; what are the advantages of Frites over other packages? If there are no other packages, simply state that Frites is the first of its kind.

Indeed, it was a missing information in the previous manuscript. We now added the sentence : *To the best of our knowledge, `Frites` is the only Python software combining information-theoretical approaches with flexible group-level analyses for the discovery of cognitive brain networks.*

## Quality of writing

> The authors do not struggle with the tenets of structural composition. The central message of the statement of need is not currently clear, but this is not an issue with the sentence-by-sentence structure/language/quality per se.

We hope that this new version clarifies the *Statement of need*.

# Functionality Review

## Functionality

> On the [main documentation page](https://brainets.github.io/frites/index.html) it might be natural to have a section called something like "What Frites can do for you" between "Description" and "Highlights" which lists the functionality claims and direct link to a corresponding script from the examples section.

We thank the reviewer for the in-depth analysis of the functionality of Frites and proposal for amelioration. As suggested, we added a section in the[ main documentation page ](https://brainets.github.io/frites/index.html) entitled [“What can you do with Frites?”](https://brainets.github.io/frites/index.html#what-can-you-do-with-frites) between "Description" and "Highlights" which lists the functionality claims and direct link to a corresponding overview and script from the examples section.

## Performance

> The claim here appears to be that the design of Frites is optimal when considering the task of analyzing particular types of data sets. What does optimal mean here and where can evidence be found for this claim? I did not see, for instance, a documentation example where Frites is compared to other software to show that it is "optimal" amongst the current choices.

We removed the term “optimal” from the paragraph. Our intention was not to make any claim about optimality, but rather to suitability. The new paragraph now reads as follows:

*Frites is suitable for the analysis of continuous and multi-channel neurophysiological data, encompassing recordings with either uniform spatial sampling (e.g., M/EEG data) and spatially sparse intracranial recordings, such as intracranial EEG or Local Field Potentials (LFPs).*

> Similar to above. Where do I find any real information about what workflows really are, let alone how they are automated and enable more efficiency? I assume that would be [here](https://brainets.github.io/frites/overview/ovw_frites.html#main-frites-workflows) but this page only has a cursory discussion of workflows and leaves me with no real concept of what they actually are beyond whatever my intuition suggests from the name "workflow". There is an "in a nutshell" discussion which is not particularly helpful in addition to a link to "checkout those examples" which takes me to the main examples page so that I must dig to find examples that demonstrate the workflows themselves.

We do agree with the reviewer that the description of the WorkFlows was not sufficient. We now added a clearer description of the workflows in the section [Step by step presentation of the workflow](https://brainets.github.io/frites/overview/ovw_frites.html#step-by-step-presentation-of-the-workflow).

In addition, we removed any claim of efficient usage from the text to avoid any mis-understanding. In fact, we meant that workflows are higher-level scripts that can be used to run multiple analysis steps.

> Moreover, the links to frites.dataset, frites.workflow and frites.conn on the [overview page](https://brainets.github.io/frites/overview/ovw_frites.html) (above the workflow section) all end in a 404.

The dead links on the [overview page](https://brainets.github.io/frites/overview/ovw_frites.html) should now all be working as expected.

> Which computations are the most demanding? Can we point to the examples directly in addition to any performance gains from using Joblib? This would fit well on a "Performance" documentation page as mentioned above

>> *In addition, some functions can further be accelerated using the Numba compiler as an optional dependency. Programming optimizations and external dependencies allow to investigate large-scale datasets in a reasonable time.*

> What functions can be accelerated / where is acceleration most profitable in Frites? To justify the claim of "large scale datasets in reasonable time" it seems natural to include a benchmark plot showing the processing time of Frites, versus other software packages as well, against a data set size.

We thank the reviewer for this great idea of illustrating the performance of Frites. We therefore included a [Performance](https://brainets.github.io/frites/overview/ovw_perf.html) page on the website, in the overview section. This section explains that we are mainly using three tricks to decrease computing time :
* Tensor computations illustrated [here](https://brainets.github.io/frites/auto_examples/performance/plot_vector_vs_tensor.html#sphx-glr-auto-examples-performance-plot-vector-vs-tensor-py)
* Parallel computing. This one is not trivial to show live in the documentation, mainly because gains appear with larger data and even more on computer clusters, which is not feasible on the documentation. Also, an example running on a laptop could potentially led to very different results depending on the OS and materials.
* Code compilation : we specifically pointed toward the functions that can be compiled if Numba is installed.

We want to thank again @travisbthomp for this detailed review. We hope that those revisions are going to clarify the need for this software, even for non-expert.
