### 2022-07-12 984bb82
Add the possible links with IDTxl, infotheory and BCT
