+++ paper/paper.bib
@@ -180,3 +180,20 @@
     pages = {177--190},
 }
 
+@article{Rubinov:2010,
+    title = {Complex network measures of brain connectivity: uses and interpretations},
+    volume = {52},
+    number = {3},
+    journal = {Neuroimage},
+    author = {Rubinov, Mikail and Sporns, Olaf},
+    year = {2010},
+    note = {Publisher: Elsevier},
+    pages = {1059--1069},
+}
+
+@article{Candadai:2019,
+  title={infotheory: A c++/python package for multivariate information theoretic analysis},
+  author={Candadai, Madhavun and Izquierdo, Eduardo J},
+  journal={arXiv preprint arXiv:1907.02339},
+  year={2019}
+}
