### 2022-07-12 880802c
Add sentence for a broader application of frites
