+++ paper/paper.md
@@ -53,7 +53,7 @@ The field of cognitive computational neuroscience addresses open questions regar
 `Frites` is suitable for the analysis of continuous and multi-channel neurophysiological data, encompassing non-invasive recordings with uniform spatial sampling (e.g., M/EEG data) such as invasive and spatially sparse recordings, such as intracranial EEG or Local Field Potentials (LFPs). The package supports standard [`NumPy`](https://numpy.org/) array inputs [@Harris:2020], neuro-oriented objects from the [`MNE-Python`](https://mne.tools/stable/index.html) software [@Gramfort:2013], but also multi-dimensional labelled  [`Xarray`](http://xarray.pydata.org/en/stable/) objects [@Hoyer:2017].
 
 ## Measures of information and group-level statistics
-`Frites` is equipped with a set of information theoretic tools for the analysis of interactions between brain signals and their relation with experimental task-related variables. By default, `Frites` is using the Gaussian Copula Mutual-Information [@Ince:2017] to study the relation between either local brain activity, cross-frequency coupling [@Combrisson:2020] or inter-areal FC with experimental variables (i.e., cognitive tasks). For what concerns FC, the toolbox allows the estimate of dynamic (i.e., time-resolve) [@Brovelli:2017], undirected (e.g., mutual information) and directed (e.g., Granger causality) FC on a single-trial basis [@Brovelli:2015]. The networks built from the FC can then be further analyzed using graph-theoretical approaches from the `Brain Connectivity Toolbox` [Rubinov:2010]. Nevertheless, `Frites` is not limited to the included estimators as the definition of custom ones is also supported, such as [`IDTxl`](https://github.com/pwollstadt/IDTxl) kernel methods [@Wollstadt:2019], [`infotheory`](https://github.com/madvn/infotheory) information decomposition [@Candadai:2019] or [`scikit-learn`](https://scikit-learn.org/stable/) cross-validated classifiers [@Pedregosa:2011]. The estimated information can then be combined with the group-level statistics for the identification of feature-specific networks. While `Frites` is strongly oriented toward neuroscience applications, the estimators of information are agnostic to data types and can therefore be applied to different fields.
+`Frites` is equipped with a set of information theoretic tools for the analysis of interactions between brain signals and their relation with experimental task-related variables. By default, `Frites` is using the Gaussian Copula Mutual-Information [@Ince:2017] to study the relation between either local brain activity, cross-frequency coupling [@Combrisson:2020] or inter-areal FC with experimental variables (i.e., cognitive tasks). For what concerns FC, the toolbox allows the estimate of dynamic (i.e., time-resolve) [@Brovelli:2017], undirected (e.g., mutual information) and directed (e.g., Granger causality) FC on a single-trial basis [@Brovelli:2015]. The networks built from the FC can then be further analyzed using graph-theoretical approaches from the `Brain Connectivity Toolbox` [@Rubinov:2010]. Nevertheless, `Frites` is not limited to the included estimators as the definition of custom ones is also supported, such as [`IDTxl`](https://github.com/pwollstadt/IDTxl) kernel methods [@Wollstadt:2019], [`infotheory`](https://github.com/madvn/infotheory) information decomposition [@Candadai:2019] or [`scikit-learn`](https://scikit-learn.org/stable/) cross-validated classifiers [@Pedregosa:2011]. The estimated information can then be combined with the group-level statistics for the identification of feature-specific networks. While `Frites` is strongly oriented toward neuroscience applications, the estimators of information are agnostic to data types and can therefore be applied to different fields.
 
[…]
