### 2022-07-12 #3842 @EtienneCmb
Dear @madvn,
> Looks like a very interesting toolbox. I think the idea of feature-specific functional networks using mutual information can be used widely in not just neuroscientific settings but also sociology or even, more topically, epidemiology. I've updated my checklist based on a quick look at the repo and installing the package. I also took a quick look at the paper and have some notes for the authors based on the following requirements while I work my way through the rest of the checklist.

We thank the reviewer for this great remark and for the positive comments. We will try to address the concerns raised.

> You have a summary but please revise it to address the "for a diverse, non-specialist audience" part of the requirement. Add a note about the breadth of usability of these tools in other domains besides neuroscience.

We agree that the first version of the “Summary” section was hard to grasp for non-expert readers. We simplified this section and also clarified why measures of information are important for brain data analyses. About the breadth of usability, we added the following sentence to the `Measures of information and group-level statistics` section :

*While Frites is strongly oriented toward neuroscience applications, the estimators of information are agnostic to data types and can therefore be applied to different fields.*

> Please place your tool in the space of other information theory tools such as infotheory, IDTXL etc. as well as network neuroscience tools such as Brain connectivity toolbox etc. I do believe there are unique aspects to your package, such as combining information-theoretic measures with network neuroscience to find feature-specific functional networks (I'm not aware of other tools that does this -- I will look it up. If you are aware of others, please include those.

We thank the reviewer for pointing us toward `infotheory` and the possible link between Frites and the `BCT`. We added the following sentence to the revised manuscript for encompassing those packages :

*The networks built from the FC can then be further analyzed using graph-theoretical approaches from the [`Brain Connectivity Toolbox`](https://sites.google.com/site/bctnet/) (Rubinov and Sporns, 2010). Nevertheless, `Frites` is not limited to the included estimators as the definition of custom ones is also supported, such as [`IDTxl`](https://github.com/pwollstadt/IDTxl) kernel methods (Wollstadt et al, 2019), [`infotheory`](https://github.com/madvn/infotheory) information decomposition (Candadai & Izquierdo, 2019) or [`scikit-learn`](https://scikit-learn.org/stable/) cross-validated classifiers (Pedregosa et al., 2011). The estimated information can then be combined with the group-level statistics for the identification of feature-specific networks. While `Frites` is strongly oriented toward neuroscience applications, the estimators of information are agnostic to data types and can therefore be applied to different fields.*

We sincerely thank the reviewer for the feedback,
