### 2022-07-31 #3842 @travisbthomp
Dear @EtienneCmb  @brovelli (CC: @meg-simula),

Thank you for making changes pursuant to the review of Frites.  Regarding advancing the reviewing process (@brovelli):  I am currently traveling overseas; I will be happy to review the comments left by @EtienneCmb when I return in a few weeks.
