### 2022-09-07 #3842 @travisbthomp
@EtienneCmb @brovelli @meg-simula

I have completed my review of Frites for this round.  I read through the changes and I wanted to thank both  @EtienneCmb @brovelli for their work addressing readability issues in addition to severe documentation shortfalls (such as dead links and examples that did not run, etc).  I also noted that the authors re-evaluated their performance claims and added a "Performance" documentation page, which is a nice addition since it draws the readers' attention to Joblib and Numba.  The reviewer's also clarified the current state of the art and Frites' contextual position within this area.

Due to the author's efforts, I have now checked 14 / 19 of the review checklist items.  Below, I have detailed each item in the checklist that was checked this round or left unchecked.  Each item is marked as **Resolved** (checked) or includes a **To Resolve** suggestion.  One item is marked **Resolved (with a suggestion)** indicating that it has been checked but that the authors might consider / reflect upon the suggestion at their discretion.

**Review Checklist Items**

* **General Checks**
    1. > Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)

    **Resolved** I believe that Frites meets the bar for a substantial scholarly effort.

* **Functionality**

    1. > Have the functional claims of the software been confirmed?

    Partially.  The authors have demonstrated many of the capabilities of the software in their examples section.  However, I have raised an issue with one of the two explicitly stated goals of the software.  As the functional claim verification occurs in the documentation as part of the examples, I defer this comment to item (6) of the next section

    **To resolve**: See item (6) of the next section.

    2. >  If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

    **Resolved** The authors addressed this in  their review response.  They have removed the performance claims that were brought up in the first round of review.

* **Documentation**
    1. > Do the authors clearly state what problems the software is designed to solve and who the target audience is?

    The **Overview** page of the documentation clearly states two goals for Frites.  No target audience is stated in the documentation.  **To Resolve**: Please add a target audience statement to the documentation.

    3. > Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

    **Resolved** Yes.  The authors provide an installation main page that lists all dependencies and includes instructions for installation with Pip.

    4. > Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

    **Resolved** Many examples are included.
 These are mostly synthetic, but it is understandable that the authors may not have access / permission to release raw data sets.  The authors might consider illustrating Frites on some freely available EEG data or similar.

    5. > Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

    **Resolved (with a suggestion)** I have marked this as resolved as the authors have included enough examples (in the examples section) to meet the very barest of minimal requirements as the examples do contain the API objects if one digs through them.  I do have a **suggestion**, though.  In the "API" section of the documentation you list several API objects/items (Dataset, Workflows, Connectivity, etc).  Clicking on these items (e.g. the frites.workflow link) takes me to a separate page.  It would be useful to, at the very least, **include links to the examples** that use each cited API (e.g. WfMi, WfMiCombine, WfConnComod, etc) object on this landing page.  In this way, the users can at least see a demonstration of these various API and get some minimal context for how they are used.  The API documentation pages are currently as sparse as possible while still being minimally feasible for publication.  If the authors want others to actually use their software, post-publication, the API documentation could be made more verbose.  This is especially true given the author's explicit claim to be targeting "researchers and students with little programming knowledge".

    6. > Are there automated tests or manual steps described so that the functionality of the software can be verified?

    The Frites documentation currently ships with several examples.  As far as I can see, these currently serve as the "automated tests" that allow for verification of the software functionality.  The "Goals" page lists two goals of Frites: (A) Extract task-related cognitive brain networks and (B) Group-level statistical inferences.  As a reviewer, I expect the main goals for the functionality to have verifiable examples / resources available.  One can argue that (B) is verifiable as there are seven examples of statistical tests in the "Examples" section (for two variables only).  However, I cannot find anywhere where a task-related cognitive **brain network** is extracted / produced.  I do see that the authors have included "Autoregressive Models" examples.  I looked through these and they do show how Granger causality can be used to infer connectivity but they use synthetic networks a priori.  From the documentation goals, one would expect a set of a-priori signal data to be used to create a graph from the GC analysis.  I suspect that this would be one of the main tasks that a user of Frites would want to carry out.

    **To Resolve**: To resolve this issue, a clear verification of the author's stated goal (A) seems to be needed.  One thing that I can think of would be to create and release a data set (using, e.g. StimSpecAR) of node-wise signals, reflecting say an ~8-10 node graph, and create and example that imports the nodal stimulus data set and then re-creates the connectivity and outputs the weighted graph (or correlation matrix or both) from those signals.  The authors may have a better idea.

    7. >  Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

    I do not see any anything explicit in this direction on the online documentation.

    **To resolve**: Add an explicit mention of this item in line with the review checklist requirements.

* **Software Paper**

    1. > Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

    The authors revised and simplified the summary section for readability.  I think this section is readable by a non-specialist audience.  However, this section lacks "a clear description of the high-level functionality and purpose of the software".  One could argue that such a description of purpose is provided: (a) implicitly by the final sentence about the "[need for] powerful measures of information to detect complex relation**s** and statistical relevance" (but this is vague) or (b) in the statement of need (but this is not in the summary section, as required by the guidelines)

  **To resolve**: please include "a clear description of the high-level functionality and purpose of the software" in the summary section as requested by the checklist.

    2. >Statement of need: Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?

    **Resolved** The statement of need now includes a target audience statement in addition to stating that [Frites is for those] working on the discovery of cognitive brain networks using neurophysiological data.  The high-level functionality discussed are: information based analysis, group-level statistical inference, functional connectivity measurement.  In light of these observations, I would again cite the need for the authors to demonstrate (e.g. with an example in the documentation) how to "**construct** a cognitive brain **network** using neurophysiological data".
