### 2022-09-16 70fa02d
Issue and PR template + description for developers in the doc
