+++ docs/source/overview/ovw_help.rst
@@ -1,4 +1,17 @@
-Get help
-========
+Report, contribute and get help
+===============================
 
-If you have questions or if you want to start discussing about new features and ideas, use the `Discussion <https://github.com/brainets/frites/discussions>`_ tab directly from the Github repository of Frites.
+Report an issue with the software
++++++++++++++++++++++++++++++++++
+
+For reporting a bug about the software (i.e. bug report, documentation or feature request), please submit an `issue <https://github.com/brainets/frites/issues/new/choose>`_.
+
+Contribute to the software
+++++++++++++++++++++++++++
+
+For developers aiming to contribute to the software, please submit your `pull request <https://github.com/brainets/frites/compare>`_.
+
+Get help
+++++++++
+
+If you have questions use the `Discussion <https://github.com/brainets/frites/discussions>`_ tab directly from the Github repository of Frites.
