+++ .github/ISSUE_TEMPLATE/documentation.yml
@@ -1,0 +1,24 @@
+name: 📄 Documentation
+
+# See the json schema to help you create / update this form
+# https://json.schemastore.org/github-issue-forms.json
+
+description: Tell us about tutorials you wish we had in our docs, ideas for improving existing tutorials, terms missing from our glossary, or parts of our docstrings that are confusing or inaccurate.
+
+labels: ["DOC"]
+
+body:
+
+  - type: markdown
+    attributes:
+      value: |
+        Please make the issue title a succinct description of your proposed change.
+
+  - type: textarea
+    attributes:
+      label: Proposed documentation enhancement
+      description: |
+        Describe your proposed enhancement in detail. If you are requesting a new
+        glossary term, please include a proposed definition.
+    validations:
+      required: true
