### 2022-10-03 b5443ac
Use linear sliding windows instead of closest time point
