+++ .github/workflows/test_doc.yml
@@ -50,6 +50,7 @@ jobs:
       - name: Build the Doc 🔧
         run: |
           cd docs
+          make clean
           make html
           touch build/html/.nojekyll
 
