+++ .github/workflows/test_doc.yml
@@ -40,7 +40,7 @@ jobs:
       - name: Set up Python 🔧
         uses: actions/setup-python@v4
         with:
-          python-version: '3.7'
+          python-version: '3.9'
 
       - name: Install dependencies 🔧
         run: |
