### 2022-10-04 a480371
Add heatmap plotting function
