+++ frites/plot/plot_conn.py
@@ -1,0 +1,192 @@
+import numpy as np
+import xarray as xr
+import pandas as pd
+
+from frites.conn import conn_reshape_undirected
+
+
+def _prepare_plot_conn(
+        conn, cmap=None, bad=None, vmin=None, vmax=None, categories=None,
+        ax=None
+    ):
+    """Prepare inputs."""
+    import matplotlib.pyplot as plt
+
+    cfg = dict()
+
+    # __________________________________ I/O __________________________________
+    # connectivity array
+    if isinstance(conn, np.ndarray):
+        assert conn.ndim == 2
+        n_rows, n_cols = conn.shape
+        roi = np.arange(conn.shape[0]).astype(str)
+        conn = pd.DataFrame(conn, index=np.arange(n_rows),
+                            columns=np.arange(n_cols))
+    elif isinstance(conn, xr.DataArray):
+        assert conn.ndim == 2
+        conn = conn.to_pandas()
+    assert isinstance(conn, pd.DataFrame)
+    np.testing.assert_array_equal(conn.index, conn.columns)
+    conn.index = conn.columns = [str(k) for k in conn.index]
+
+    # _________________________________ NODES _________________________________
+    # _________________________________ COLOR _________________________________
+    # colormap
+    cmap = plt.get_cmap(cmap).copy()
+    if bad:
+        cmap.set_bad(color=bad)
+    cfg['cmap'] = cmap
+
+    # vmin, vmax trick
+    if isinstance(vmin, str):
+        vmin = np.nanpercentile(conn.values, float(vmin))
+    if isinstance(vmax, str):
+        vmax = np.nanpercentile(conn.values, float(vmax))
+    cfg['vmin'], cfg['vmax'] = vmin, vmax
+
+    # ______________________________ CATEGORIES _______________________________
+    cfg['has_categories'] = False
+    if isinstance(categories, (list, np.ndarray, tuple)):
+        cat_cut = np.diff(np.unique(categories, return_inverse=True)[1]) != 0
+        cut_at = np.where(cat_cut)[0] + 1
+        cfg['has_categories'] = True
+        cfg['categories'] = categories
+        cfg['cut_at'] = cut_at.astype(int)
+
+    # ________________________________ FIGURE _________________________________
+    if ax is None:
+        cfg['fig'] = plt.figure(figsize=(14, 6))
+        cfg['ax'] = plt.gca()
+    else:
+        cfg['fig'] = plt.gcf()
+        cfg['ax'] = plt.gca()
+        plt.sca(ax)
+
+    return conn, cfg
+
+
+###############################################################################
+###############################################################################
+#                                 HEATMAP
+###############################################################################
+###############################################################################
+
+
+def plot_conn_heatmap(
+        conn, cmap='plasma', vmin=None, vmax=None, categories=None,
+        categories_kw={}, cbar=True, cbar_title=None, cbar_kw={},
+        bad=None, xticklabels='auto', yticklabels='auto', square=True, ax=None
+    ):
+    """Plot the connectivity matrix as a heatmap.
+
+    Parameters
+    ----------
+    conn : xarray.DataArray | pandas.DataFrame | numpy.ndarray
+        Either a 2D xarray.DataArray or a pandas DataFrame or a 2D NumPy array
+    cmap : str | 'plasma'
+        Colormap name
+    vmin, vmax : float | None
+        Minimum and maximum of colorbar limits
+    categories : array_like | None
+        Category associated to each region name. Can be hemisphere name,
+        lobe name or indices describing group of regions. By default, an
+        horizontal and a vertical lines are going to be plotted as a separation
+        between categories (see argument below for controlling the aesthetic)
[…]
