+++ frites/utils/__init__.py
@@ -2,6 +2,6 @@
 from .parallel import parallel_func  # noqa
 from .preproc import (  # noqa
     savgol_filter, kernel_smoothing, downsample, acf, nonsorted_unique,
-    time_to_sample, get_closest_sample
+    time_to_sample, get_closest_sample, normalize
 )
 from .wrapper import jit  # noqa
