+++ frites/plot/plot_conn.py
@@ -1,15 +1,18 @@
+"""Plot connectivity matrix.
+"""
 import numpy as np
 import xarray as xr
 import pandas as pd
 
-from frites.conn import conn_reshape_undirected
+from frites.utils import normalize
 
 
 def _prepare_plot_conn(
         conn, cmap=None, bad=None, vmin=None, vmax=None, categories=None,
-        ax=None
-    ):
+        nodes_data=None, nodes_cmap=None, nodes_bad=None, ax=None,
+        square=False, prop=None, polar=False):
     """Prepare inputs."""
+    import matplotlib as mpl
     import matplotlib.pyplot as plt
 
     cfg = dict()
@@ -19,7 +22,6 @@ def _prepare_plot_conn(
     if isinstance(conn, np.ndarray):
         assert conn.ndim == 2
         n_rows, n_cols = conn.shape
-        roi = np.arange(conn.shape[0]).astype(str)
         conn = pd.DataFrame(conn, index=np.arange(n_rows),
                             columns=np.arange(n_cols))
     elif isinstance(conn, xr.DataArray):
@@ -28,8 +30,45 @@ def _prepare_plot_conn(
     assert isinstance(conn, pd.DataFrame)
     np.testing.assert_array_equal(conn.index, conn.columns)
     conn.index = conn.columns = [str(k) for k in conn.index]
+    n_nodes = len(conn.index)
+
+    # proportion of edges to keep
+    if isinstance(prop, (int, float)):
+        assert 0 < prop < 100
+        th = np.nanpercentile(conn.values, 100. - prop)
+        conn[conn < th] = np.nan
 
     # _________________________________ NODES _________________________________
+    # nodes name and order
+    cfg['nodes_names'] = conn.index.tolist()
+    cfg['nodes_order'] = conn.index.tolist()
+
+    # nodes data
+    if isinstance(nodes_data, (list, np.ndarray, tuple)):
+        if len(nodes_data) != n_nodes:
+            raise ValueError(
+                "When passing custom data to the nodes, the `nodes_data` must "
+                "have the same length as the number of nodes")
+        nodes_data = np.asarray(nodes_data)
+    elif nodes_data == 'degree':
+        nodes_data = (np.isfinite(conn.values)).sum(axis=0).astype(int)
+    elif nodes_data == 'mean':
+        nodes_data = np.nanmean(conn.values, axis=0)
+    elif nodes_data == 'number':
+        nodes_data = np.arange(n_nodes)
+    else:
+        nodes_data = np.full((n_nodes,), np.nan)
+    nodes_data = normalize(nodes_data, to_min=0., to_max=1.)
+    cfg['nodes_data'] = np.ma.masked_array(
+        nodes_data, mask=~np.isfinite(nodes_data)
+    )
+
+    # nodes color
+    nodes_cmap = plt.get_cmap(nodes_cmap).copy()
+    if nodes_bad:
+        nodes_cmap.set_bad(color=nodes_bad)
+    cfg['nodes_color'] = [nodes_cmap(k) for k in cfg['nodes_data']]
+
     # _________________________________ COLOR _________________________________
     # colormap
     cmap = plt.get_cmap(cmap).copy()
@@ -42,7 +81,11 @@ def _prepare_plot_conn(
         vmin = np.nanpercentile(conn.values, float(vmin))
     if isinstance(vmax, str):
         vmax = np.nanpercentile(conn.values, float(vmax))
-    cfg['vmin'], cfg['vmax'] = vmin, vmax
+    if not isinstance(vmin, (int, float)):
+        vmin = np.nanmin(conn.values)
+    if not isinstance(vmax, (int, float)):
+        vmax = np.nanmax(conn.values)
[…]
