### 2022-10-05 24029c4
Make tests and doc for plotting functions
