+++ frites/plot/tests/test_plot_conn.py
@@ -1,0 +1,137 @@
+"""Test plotting connectivity."""
+import numpy as np
+import xarray as xr
+import pandas as pd
+
+from frites.plot.plot_conn import (
+    _prepare_plot_conn, plot_conn_heatmap, plot_conn_circle
+)
+
+
+class TestPlotConn(object):
+
+    @staticmethod
+    def _get_conn(astype='numpy'):
+        conn = [
+            [np.nan, 2, 1],
+            [2, 2, 4],
+            [-1, np.nan, 1]
+        ]
+        roi = ["roi_0", "roi_1", "roi_2"]
+        conn = np.asarray(conn)
+
+        if astype == 'numpy':
+            return conn
+        elif astype == 'pandas':
+            return pd.DataFrame(conn, index=roi, columns=roi)
+        elif astype == 'xarray':
+            return xr.DataArray(conn, dims=('sources', 'targets'),
+                                coords=(roi, roi))
+
+    def test_prepare_inputs(self):
+        """Test input transformation for connectivity plots."""
+        import matplotlib.pyplot as plt
+        from matplotlib.colors import ListedColormap
+
+        plt.figure()
+
+        # test input types
+        for types in ['numpy', 'pandas', 'xarray']:
+            conn = self._get_conn(astype=types)
+            out = _prepare_plot_conn(conn)[0]
+            assert isinstance(out, pd.DataFrame)
+
+        # test color settings of edges
+        conn = self._get_conn(astype='pandas')
+        out = _prepare_plot_conn(conn, cmap='Set1', bad='orange')[1]
+        assert isinstance(out['cmap'], ListedColormap)
+        assert (out['vmin'] == -1.) and (out['vmax'] == 4.)
+        out = _prepare_plot_conn(conn, vmin='20', vmax='80')[1]
+        assert (out['vmin'] == 1.) and (out['vmax'] == 2.)
+        plt.close()
+
+        # test categories
+        conn = self._get_conn(astype='pandas')
+        out = _prepare_plot_conn(conn, categories=[0, 0, 1])[1]
+        np.testing.assert_array_equal(out['categories'], [0, 0, 1])
+        plt.close()
+
+        # test nodes data
+        conn = self._get_conn(astype='pandas')
+        out = _prepare_plot_conn(conn, nodes_data=None)[1]
+        np.testing.assert_array_equal(out['nodes_data'], [1.] * 3)
+        out = _prepare_plot_conn(conn, nodes_data='degree')[1]
+        np.testing.assert_array_equal(out['nodes_data'], [0, 0, 1])
+        out = _prepare_plot_conn(conn, nodes_data='mean')[1]
+        np.testing.assert_array_equal(out['nodes_data'], [0, 1, 1])
+        out = _prepare_plot_conn(conn, nodes_data='diagonal')[1]
+        np.testing.assert_array_equal(out['nodes_data'], [np.nan, 1, 0])
+        out = _prepare_plot_conn(conn, nodes_data='number')[1]
+        np.testing.assert_array_equal(out['nodes_data'], [0, .5, 1.])
+        out = _prepare_plot_conn(conn, nodes_data='oups')[1]
+        np.testing.assert_array_equal(out['nodes_data'], [1] * 3)
+        out = _prepare_plot_conn(conn, nodes_data=[3, 2, 1])[1]
+        np.testing.assert_array_equal(out['nodes_data'], [1, .5, 0])
+        plt.close()
+
+        # test nodes coloring
+        conn = self._get_conn(astype='pandas')
+        out = _prepare_plot_conn(conn, nodes_cmap='Set1', nodes_bad='k')[1]
+        plt.close()
+
+        # test axis definition
[…]
