+++ docs/source/api/api_plot.rst
@@ -1,0 +1,19 @@
+Information-based estimators
+----------------------------
+
+:py:mod:`frites.plot`:
+
+.. currentmodule:: frites.plot
+
+.. automodule:: frites.plot
+   :no-members:
+   :no-inherited-members:
+
+Plotting functions for connectivity estimations
++++++++++++++++++++++++++++++++++++++++++++++++
+
+.. autosummary::
+   :toctree: generated/
+
+   plot_conn_heatmap
+   plot_conn_circle
