+++ frites/plot/plot_conn.py
@@ -7,10 +7,41 @@ import pandas as pd
 from frites.utils import normalize
 
 
+def _compute_nodes_data(conn, nodes_data, to_min=0., to_max=1.):
+    """Compute data to attach to each node."""
+    n_nodes = conn.values.shape[0]
+
+    if isinstance(nodes_data, (list, np.ndarray, tuple)):
+        if len(nodes_data) != n_nodes:
+            raise ValueError(
+                "When passing custom data to the nodes, the `nodes_data` must "
+                "have the same length as the number of nodes")
+        nodes_data = np.asarray(nodes_data)
+    elif nodes_data == 'degree':
+        nodes_data = (np.isfinite(conn.values)).sum(axis=0).astype(int)
+    elif nodes_data == 'mean':
+        nodes_data = np.nanmean(conn.values, axis=0)
+    elif nodes_data == 'diagonal':
+        nodes_data = np.diag(conn.values)
+    elif nodes_data == 'number':
+        nodes_data = np.arange(n_nodes)
+    else:
+        nodes_data = np.full((n_nodes,), 1.)
+
+    # normalize the data
+    nodes_data = normalize(nodes_data, to_min=to_min, to_max=to_max)
+
+    # mask non-finite values
+    nodes_data = np.ma.masked_array(nodes_data, mask=~np.isfinite(nodes_data))
+
+    return nodes_data
+
+
 def _prepare_plot_conn(
         conn, cmap=None, bad=None, vmin=None, vmax=None, categories=None,
-        nodes_data=None, nodes_cmap=None, nodes_bad=None, ax=None,
-        square=False, prop=None, polar=False):
+        nodes_data=None, nodes_size=None, nodes_size_min=None,
+        nodes_size_max=None, nodes_cmap=None, nodes_bad=None,
+        ax=None, square=False, prop=None, polar=False):
     """Prepare inputs."""
     import matplotlib as mpl
     import matplotlib.pyplot as plt
@@ -46,26 +77,12 @@ def _prepare_plot_conn(
     cfg['nodes_order'] = conn.index.tolist()
 
     # nodes data
-    if isinstance(nodes_data, (list, np.ndarray, tuple)):
-        if len(nodes_data) != n_nodes:
-            raise ValueError(
-                "When passing custom data to the nodes, the `nodes_data` must "
-                "have the same length as the number of nodes")
-        nodes_data = np.asarray(nodes_data)
-    elif nodes_data == 'degree':
-        nodes_data = (np.isfinite(conn.values)).sum(axis=0).astype(int)
-    elif nodes_data == 'mean':
-        nodes_data = np.nanmean(conn.values, axis=0)
-    elif nodes_data == 'diagonal':
-        nodes_data = np.diag(conn.values)
-    elif nodes_data == 'number':
-        nodes_data = np.arange(n_nodes)
-    else:
-        nodes_data = np.full((n_nodes,), 1.)
-    nodes_data = normalize(nodes_data, to_min=0., to_max=1.)
-    cfg['nodes_data'] = np.ma.masked_array(
-        nodes_data, mask=~np.isfinite(nodes_data)
-    )
+    cfg['nodes_data'] = _compute_nodes_data(
+        conn, nodes_data, to_min=0., to_max=1.)
+
+    # nodes_size
+    cfg['nodes_size'] =  _compute_nodes_data(
+        conn, nodes_size, to_min=nodes_size_min, to_max=nodes_size_max)
 
     # nodes color
     if (nodes_cmap is None) or isinstance(nodes_cmap, str):
@@ -253,10 +270,11 @@ def plot_conn_heatmap(
 def plot_conn_circle(
         conn, directed=False, edges_cmap='hot_r', edges_vmin=None,
[…]
