### 2022-10-12 78bbb91
Support signed connectivity and improve node control
