+++ frites/plot/plot_conn.py
@@ -67,8 +67,8 @@ def _prepare_plot_conn(
     # proportion of edges to keep
     if isinstance(prop, (int, float)):
         assert 0 < prop < 100
-        th = np.nanpercentile(conn.values, 100. - prop)
-        conn[conn < th] = np.nan
+        th = np.nanpercentile(np.abs(conn.values), 100. - prop)
+        conn[np.abs(conn) < th] = np.nan
 
     # _________________________________ NODES _________________________________
     # nodes name and order
@@ -132,7 +132,7 @@ def _prepare_plot_conn(
         ax = tuple(ax)
 
         if all([k == 1 for k in ax]):
-            cfg['fig'] = plt.figure(figsize=(14, 10))
+            cfg['fig'] = plt.figure(figsize=(16, 12))
         else:
             cfg['fig'] = plt.gcf()
 
@@ -267,10 +267,11 @@ def plot_conn_heatmap(
 
 
 def plot_conn_circle(
-        conn, directed=False, edges_cmap='hot_r', edges_vmin=None,
-        edges_vmax=None, edges_lw=3., edges_alpha=1., nodes_data='degree',
-        nodes_cmap='hot_r', nodes_size=None, nodes_size_min=1.,
-        nodes_size_max=50., nodes_bad=None, nodes_fz=8, nodes_shift=0.,
+        conn, signed=False, directed=False, edges_cmap='hot_r',
+        edges_vmin=None, edges_vmax=None, edges_lw=3., edges_alpha=1.,
+        nodes_data='degree', nodes_cmap='hot_r', nodes_size=None,
+        nodes_size_min=1., nodes_size_max=50., nodes_bad=None,
+        nodes_label_fz=8, nodes_label_color=None, nodes_label_shift=0.,
         categories=None, categories_sep=3, cbar=True, cbar_title=None,
         cbar_kw={}, cbar_size=.8, cbar_pos=(.8, .4), prop=None, angle_start=90,
         angle_span=360, padding=0., ax=None):
@@ -286,6 +287,9 @@ def plot_conn_circle(
     ----------
     conn : xarray.DataArray | pandas.DataFrame | numpy.ndarray
         Either a 2D xarray.DataArray or a pandas DataFrame or a 2D NumPy array
+    signed : bool | False
+        Specifiy whether the connectivity metric is signed (e.g. pearson or
+        spearman correlation) or unsigned (e.g. mutual information, distance)
     directed : bool | False
         Specify whether it is directed connectivity (True) or undirected
         connectivity (False, default)
@@ -324,10 +328,12 @@ def plot_conn_circle(
     nodes_bad : str | None
         Color to use for bad nodes (i.e. nodes with non finite values in
         nodes_data). By default, those nodes are transparent.
-    nodes_fz : float | 8
+    nodes_label_fz : float | 8
         Font size of nodes' labels.
-    nodes_shift : float | 0.
+    nodes_label_shift : float | 0.
         Shift to apply to the labels (i.e. the nodes names)
+    nodes_label_color : array_like | list | None
+        Color to apply to each node label
     categories : array_like | None
         Category associated to each region name. Can be hemisphere name,
         lobe name or indices describing group of regions. By default, a space
@@ -389,9 +395,10 @@ def plot_conn_circle(
     # plot the connectivity
     ax = _draw_conn_circle(
         conn.values, nodes_names, angles, cfg['nodes_color'],
-        cfg['nodes_size'], nodes_fz, nodes_shift, cfg['cmap'], cfg['vmin'],
-        cfg['vmax'], edges_lw, edges_alpha, directed, cbar, cbar_title,
-        cbar_kw, cbar_size, cbar_pos, cfg['ax'], padding
[…]
