### 2022-10-12 405eafb
Fix function signature
