### 2022-10-12 b20af1a
conn_links support categorical selection
