### 2022-10-13 #3842 @EtienneCmb
Dear @meg-simula,

We don't have news from @madvn. I also tried to reach him by email without success. To summarize, he proposed [several modifications](https://github.com/openjournals/joss-reviews/issues/3842#issuecomment-980836301) and [we modified the manuscript accordingly](https://github.com/openjournals/joss-reviews/issues/3842#issuecomment-1181985736). How can we finalize the revisions?

Thank you in advance,
