### 2022-10-13 #3842 @madvn
Hi everyone, so sorry about the delay. I'm in vacation this week and will
be back on Monday. I'll get to this Monday evening. Hope that's not too
late.

On Thu, Oct 13, 2022 at 10:09 AM Etienne Combrisson <
***@***.***> wrote:

> Dear @meg-simula <https://github.com/meg-simula>,
>
> We don't have news from @madvn <https://github.com/madvn>. I also tried
> to reach him by email without success. To summarize, he proposed several
> modifications
> <https://github.com/openjournals/joss-reviews/issues/3842#issuecomment-980836301>
> and we modified the manuscript accordingly
> <https://github.com/openjournals/joss-reviews/issues/3842#issuecomment-1181985736>.
> How can we finalize the revisions?
>
> Thank you in advance,
>
