### 2022-10-18 1250f93
Drop python 3.7 from PR testing
