### 2022-10-25 10966a1
average over tapers once connectivity metric is computed
