+++ frites/conn/conn_tf.py
@@ -66,25 +66,19 @@ def _tf_decomp(data, sf, freqs, mode='morlet', n_cycles=7.0, mt_bandwidth=None,
         # the MT decomposition is done separatedly for each
         # Frequency center
         if isinstance(mt_bandwidth, (list, tuple, np.ndarray)):
-            # Arrays freqs, n_cycles, mt_bandwidth should have the same size
-            assert len(freqs) == len(n_cycles) == len(mt_bandwidth)
-            out = []
-            for f_c, n_c, mt in zip(freqs, n_cycles, mt_bandwidth):
-                _out = tfr_array_multitaper(
-                    data, sf, [f_c], n_cycles=float(n_c), time_bandwidth=mt,
-                    output='complex', decim=decim, n_jobs=n_jobs, **kw_mt
-                )
+            raise NotImplementedError("Not compatible with multiple bandwidth")
+            # # Arrays freqs, n_cycles, mt_bandwidth should have the same size
+            # assert len(freqs) == len(n_cycles) == len(mt_bandwidth)
+            # out = []
+            # for f_c, n_c, mt in zip(freqs, n_cycles, mt_bandwidth):
+            #     _out = tfr_array_multitaper(
+            #         data, sf, [f_c], n_cycles=float(n_c), time_bandwidth=mt,
+            #         output='complex', decim=decim, n_jobs=n_jobs, **kw_mt
+            #     )
+            #     out.append(_out)
 
-                # recent version of mne allows to return the TF decomposition
-                # with the additional n_tapers dimension. This patch takes the
-                # mean over the tapers
-                if _out.ndim == 5:
-                    _out = _out.mean(2)
-
-                out.append(_out)  #<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<
-
-            # stack everything
-            out = np.stack(out, axis=2).squeeze()
+            # # stack everything
+            # out = np.concatenate(out, axis=2)
         elif isinstance(mt_bandwidth, (type(None), int, float)):
             out = tfr_array_multitaper(
                 data, sf, freqs, n_cycles=n_cycles,
