### 2022-11-02 #3842 @madvn
Thank you for the responses and changes with regards to my comments, @EtienneCmb. All items I have noted have been resolved [here](https://github.com/openjournals/joss-reviews/issues/3842#issuecomment-1181985736).

I have been able to run some of the examples on my Ubuntu machine with python 3.6. The examples and documentation look great!

Looks like some tests are failing on the master branch. The travis build seems to be successful though. Could you please resolve that?

I've checked all boxes except the one on automated testing to capture this last item. Otherwise looks great to me!

Thanks for the patience everyone and sorry again about the delay.
