### 2022-11-02 4ad29bf
Averageover tapers
