### 2022-11-02 #3842 @EtienneCmb
Hi @madvn,

Thank you for your response,

> Looks like some tests are failing on the master branch. The travis build seems to be successful though. Could you please resolve that?

Yes, we made some changes to connectivity metrics. I solved the testing issue related to this.

Thank you,
