### 2022-11-02 #3842 @madvn
Perfect, checks all boxes. LGTM @meg-simula
