### 2022-11-03 #3842 @meg-simula
@EtienneCmb Could you comment on the references missing DOIs above? Are these expected or addressable?
