### 2022-11-03 #3842 @EtienneCmb
> @EtienneCmb I've read through the revised manuscript and find it very accessible, readable and interesting. Thank you for all of your work with the revisions, and @travisbthomp @madvn for the reviews.

Thank you @meg-simula, indeed, the comments of the reviewers helped a lot in clarifying the manuscript.

> @EtienneCmb Could you comment on the references missing DOIs above? Are these expected or addressable?

I checked all 5 doi are correct. Please find below the hyperlinks to the papers if you want to double check:

1. [xarray: ND labeled arrays and datasets in Python](https://openresearchsoftware.metajnl.com/articles/10.5334/jors.148/?utm_source=TrendMD&utm_medium=cpc&utm_campaign=Journal_of_Open_Research_Software_TrendMD_0)
2. [Functional connectivity and neuronal dynamics: insights from computational methods](https://direct.mit.edu/books/edited-volume/5456/chapter-abstract/3967023/Functional-Connectivity-and-Neuronal-Dynamics?redirectedFrom=fulltext)
3. [Characterization of Cortical Networks and Corticocortical Functional Connectivity Mediating Arbitrary Visuomotor Mapping](https://www.jneurosci.org/content/35/37/12643?utm_source=TrendMD&utm_medium=cpc&utm_campaign=JNeurosci_TrendMD_1)
4. [Complex network measures of brain connectivity: uses and interpretations](https://www.sciencedirect.com/science/article/abs/pii/S105381190901074X)
5. [infotheory: A c++/python package for multivariate information theoretic analysis](https://joss.theoj.org/papers/10.21105/joss.01609) (this one is a JOSS)

Otherwise, should I do something else to check?
