### 2022-11-03 #3842 @meg-simula
@EtienneCmb Thanks for checking - could you please also add the DOis to the paper references? (So that they are not missing DOIs anymore.)
