### 2022-11-03 #3842 @EtienneCmb
@meg-simula ah yes, exact, doi are missing in the bibtex file. I update right now
