+++ paper/paper.bib
@@ -46,6 +46,7 @@
     journal = {Journal of Open Research Software},
     author = {Hoyer, Stephan and Hamman, Joe},
     year = {2017},
+    doi = {10.5334/jors.148},
     note = {Publisher: Ubiquity Press},
 }
 
@@ -82,6 +83,7 @@
     title = {Functional connectivity and neuronal dynamics: insights from computational methods},
     journal = {The Cognitive Neurosciences},
     author = {Battaglia, Demian and Brovelli, Andrea},
+    doi = {10.7551/mitpress/11442.001.0001},
     year = {2020},
 }
 
@@ -93,6 +95,7 @@
     journal = {The Journal of Neuroscience},
     author = {Brovelli, Andrea and Chicharro, Daniel and Badier, Jean-Michel and Wang, Huifang and Jirsa, Viktor},
     year = {2015},
+    doi = {10.1523/jneurosci.4892-14.2015},
     pages = {12643--12658}
 }
 
@@ -188,6 +191,7 @@
     author = {Rubinov, Mikail and Sporns, Olaf},
     year = {2010},
     note = {Publisher: Elsevier},
+    doi = {10.1016/j.neuroimage.2009.10.003},
     pages = {1059--1069},
 }
 
@@ -195,5 +199,6 @@
   title={infotheory: A c++/python package for multivariate information theoretic analysis},
   author={Candadai, Madhavun and Izquierdo, Eduardo J},
   journal={arXiv preprint arXiv:1907.02339},
+  doi = {10.21105/joss.01609},
   year={2019}
 }
