### 2022-11-03 #3842 @EtienneCmb
- [x] Version tag [v0.4.3](https://github.com/brainets/frites/tree/v0.4.3)
- [x] doi of the archived : 10.5281/zenodo.7278278
