### 2022-11-03 2731728
Add badges and upload to conda
