+++ .github/workflows/pypi-publish.yml
@@ -9,9 +9,7 @@ on:
 
 jobs:
   deploy:
-
     runs-on: ubuntu-latest
-
     steps:
     - uses: actions/checkout@v3.0.2
     - name: Set up Python
