### 2022-11-03 #3842 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
