### 2022-11-03 501a6a3
Add conda config files
