+++ setup.py
@@ -68,5 +68,7 @@ setup(
                  'Intended Audience :: Education',
                  'Intended Audience :: Developers',
                  "Programming Language :: Python :: 3.7",
-                 "Programming Language :: Python :: 3.8"
+                 "Programming Language :: Python :: 3.8",
+                 "Programming Language :: Python :: 3.9",
+                 "Programming Language :: Python :: 3.10",
                  ])
