+++ .github/conda/meta.yaml
@@ -1,0 +1,44 @@
+{% set data = load_setup_py_data() %}
+
+package:
+  name: frites
+  version: {{ data['version'] }}
+
+source:
+  path: ..
+
+build:
+  number: 0
+  script: python setup.py install --single-version-externally-managed --record=record.txt
+
+requirements:
+  build:
+    - python
+    - numpy
+    - scipy
+    - mne
+    - joblib
+    - xarray
+    - netCDF4
+    - h5netcdf
+    - scikit-learn
+
+  run:
+    - python
+    - numpy
+    - scipy
+    - mne
+    - joblib
+    - xarray
+    - netCDF4
+    - h5netcdf
+    - scikit-learn
+
+test:
+  imports:
+    - frites
+
+about:
+  home: {{ data['url'] }}
+  license: {{ data['license'] }}
+summary: {{ data['description'] }}
