+++ .github/workflows/conda-publish.yml
@@ -12,6 +12,6 @@ jobs:
       - name: publish-to-conda
         uses: fcakyon/conda-publish-action@v1.3
         with:
-          subdir: 'conda'
+          subdir: '.github/conda'
           anacondatoken: ${{ secrets.ANACONDA_TOKEN }}
           platforms: 'win osx linux'
