+++ .github/conda/meta.yaml
@@ -13,7 +13,7 @@ build:
 
 requirements:
   build:
-    - python
+    - python>=3.8
     - numpy
     - scipy
     - mne
@@ -24,7 +24,7 @@ requirements:
     - scikit-learn
 
   run:
-    - python
+    - python>=3.8
     - numpy
     - scipy
     - mne
