### 2022-11-03 76dd435
Drop bad dict for jinja
