+++ .github/conda/meta.yaml
@@ -2,7 +2,7 @@
 
 package:
   name: frites
-  version: {{ version }}
+  version: "0.4.3"
 
 source:
   path: ..
