+++ .github/workflows/conda-publish.yml
@@ -8,10 +8,9 @@ jobs:
   publish:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v1
       - name: publish-to-conda
-        uses: fcakyon/conda-publish-action@v1.3
+        uses: maxibor/conda-package-publish-action@v1.1
         with:
           subdir: '.github/conda'
-          anacondatoken: ${{ secrets.ANACONDA_TOKEN }}
-          platforms: 'win osx linux'
+          AnacondaToken: ${{ secrets.ANACONDA_TOKEN }}
