+++ .github/workflows/conda-publish.yml
@@ -12,5 +12,5 @@ jobs:
       - name: publish-to-conda
         uses: maxibor/conda-package-publish-action@v1.1
         with:
-          subdir: '.github/conda'
+          subdir: 'conda'
           AnacondaToken: ${{ secrets.ANACONDA_TOKEN }}
