+++ .github/workflows/conda-publish.yml
@@ -8,9 +8,10 @@ jobs:
   publish:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v1
-      - name: publish-to-conda
-        uses: maxibor/conda-package-publish-action@v1.1
+      - uses: actions/checkout@v3.0.2
+      - name: Publish Conda forge
+        uses: EtienneCmb/conda-publish-action@v1.0
         with:
           subdir: 'conda'
-          AnacondaToken: ${{ secrets.ANACONDA_TOKEN }}
+          anacondatoken: ${{ secrets.ANACONDA_TOKEN }}
+          platforms: 'win osx linux'
