+++ .github/workflows/conda-publish.yml
@@ -8,7 +8,7 @@ jobs:
   publish:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v3.0.2
+      - uses: actions/checkout@v3.1.0
       - name: Publish Conda forge
         uses: EtienneCmb/conda-publish-action@v1.0
         with:
