### 2022-11-04 079f15c
Simplify conda action
