+++ .github/workflows/conda-publish.yml
@@ -10,8 +10,7 @@ jobs:
     steps:
       - uses: actions/checkout@v3.0.2
       - name: Publish Conda forge
-        uses: EtienneCmb/conda-publish-action@v1.0
+        uses: EtienneCmb/conda-publish-action@v1.1
         with:
           subdir: 'conda'
-          anacondatoken: ${{ secrets.ANACONDA_TOKEN }}
-          platforms: 'win osx linux'
+          AnacondaToken: ${{ secrets.ANACONDA_TOKEN }}
