+++ .github/workflows/conda-publish.yml
@@ -1,16 +1,0 @@
-name: publish_conda
-
-on:
-  push:
-    branches: [ master ]
-
-jobs:
-  publish:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3.1.0
-      - name: Publish Conda forge
-        uses: EtienneCmb/conda-publish-action@v1.1
-        with:
-          subdir: 'conda'
-          AnacondaToken: ${{ secrets.ANACONDA_TOKEN }}
