+++ conda/meta.yaml
@@ -1,46 +1,0 @@
-{% set data = load_setup_py_data() %}
-
-package:
-  name: frites
-  version: "0.4.3"
-
-source:
-  path: ..
-
-build:
-  number: 0
-  script: python setup.py install --single-version-externally-managed --record=record.txt
-
-requirements:
-  build:
-    - python>=3.8
-    - numpy
-    - pip
-    - scipy
-    - mne
-    - joblib
-    - xarray
-    - netCDF4
-    - h5netcdf
-    - scikit-learn
-
-  run:
-    - python>=3.8
-    - numpy
-    - scipy
-    - mne
-    - joblib
-    - xarray
-    - netCDF4
-    - h5netcdf
-    - scikit-learn
-
-test:
-  imports:
-    - frites
-
-about:
-  home: https://github.com/brainets/frites
-  license: BSD
-  license_file: LICENSE
-  summary: 'Framework for Information Theoretical analyses of Electrophysiological data and Statistics'
