### 2022-11-14 #3842 @madvn
@EtienneCmb , quick note, (maybe you've already planned this) you could put the pdf from here on arxiv for more visibility if you want.
